[roll] Roll fuchsia [virtualization] replace 'blocking::Unblock' with fdio file conversion
This CL removes blocking::Unblock and used EventedFd's to ensure the
CLI tool is only using sockets to communicate with Guests. This enables
consistency in how we both test the tool, as well as how we handle
all three data streams that are used when communicating with a Guest.

Testing: fx test guest-tests
Original-Bug: 89427
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/658023
Original-Revision: 5392684fa3a53a574dbac3a5004707893d3db288
GitOrigin-RevId: b5907f8172afe1452f2e1932e60850e48728c007
Change-Id: I48baece6de564fc3e2230e187db16eb16ff7e730