[roll] Roll fuchsia [input] Wires up the focus distribution to the input pipeline.
The focus distribution went unused for the longest of times, but should not
be unused. The reason this went unused was that Ermine session shell did its
own routing, which it should not be doing. We're now fixing the issue.

Since proper routing for focus distribution is now split across multiple
repositories, we need a soft transition to reroute the relevant FIDL protocols
from Ermine back to the input pipeline (scene_manager in this case).  This is
the first in a series of changes required to make this happen.

Multiply: experiences_ermine_terminal_e2e_test
Multiply: experiences_ermine_terminal_smoke_test
Original-Fixed: 98412
Original-Bug: 95901
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/666844
Original-Revision: 183ab971eb9e4afd84144c5b96aec7f2a000af01
GitOrigin-RevId: 1c0c290b2eca882f927e0afb732a54b83e823544
Change-Id: I99a68ba3a323ed5f2856e3284bfba6dfac7d5060