[roll] Roll fuchsia [fdf] Support directory composite routing
Add support for composite devices for our directory routing.
Service instance aggregation is not supported yet, so we cannot
use services. But we can route subdirectories from parents to
children.

A directory capability name with a '-' represents a service
instance with the pattern:

 SERVICE_NAME-INSTANCE_NAME

If we are routing to a composite device, and we see a "default"
service instance being routed from a parent, that instance will
be renamed to the composite node name.

The primary composite node will be routed a second time as the
'default' instance.

TEST: fx test composite_test
Original-Bug: 93678

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/673982
Original-Revision: 56d67e0f70a117e5db117db7fdf326c2ba47bf62
GitOrigin-RevId: 03efd367603c32b9c1d3533a4e7aee99f9629d07
Change-Id: I3029671f93df9e1b3eea5f0a0b2a73bdff058076