[roll] Roll fuchsia [starnix][binder] Improve reference counting
The reference counting implementation now correctly tracks weak and
strong references to binder objects in the handle table.

An object is kept in the handle table until all references, weak and
strong, are released. The underlying object it points to may be dead
earlier, however.

When a transaction occurs, any binder objects inside the buffer have a
strong reference taken by the kernel, and that reference is dropped when
the buffer is freed with a BC_FREE_BUFFER command.

Weak references are tracked, but they aren't communicated to the owning
process. Android doesn't support sending weak references, then upgrading
them to strong references. There was work in the kernel to support this
supposedly, but userspace never implemented it and the APIs were
deprecated and removed.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/674297
Original-Revision: 07d0e357f956d7a006b49b68fd3242ebd1cce5d6
GitOrigin-RevId: 3e46414a17937505084592096c1d6e0d0fb45b94
Change-Id: Idb905fbd1cb2a3b6a8847c8ee652f19cb7d07a91