[roll] Roll fuchsia [fdf][compat] Support fragment names
Make it simple to support fragment names in the compat shim.
We can figure out the fragment names based on the device service
instances that are given to us.

Unfortunately, we have to add some node information to the DFv1
CML file so the instances are routed. Add a GN argument for this
so it can happen easily (and so it can be easily replaced when
we move to services).

Original-Bug: 93678

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/674084
Original-Revision: 36c91ae52e9e203eb6523d7990d37989235bf47b
GitOrigin-RevId: dc9e73ce0973ca0b0539c86ffe12ce761bb0d27f
Change-Id: Ie58f7ad86cbadf7afddf6e91655dc326dc9fb8b1