[roll] Roll fuchsia [scenic] Publish a ViewTreeSnapshot after a frame is rendered.
This CL changes the behavior of view tree snapshotter to publish a
snapshot and supply it to its subscribers after a frame is rendered.
This solves the problem of duplicate snapshots which were published when
a frame gets dropped.

Original-Fixed: 95186

Test: fx lsc presubmit fxr/674003

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/674003
Original-Revision: e4e49b45c666293d9f17f23b5798a60812d35bbd
GitOrigin-RevId: bcdbb2b94c775f11a921709098bcf23435c9ebd4
Change-Id: I5b6e2a99de7c5320b7ea683d8697c1391bcd2ab0