[roll] Roll fuchsia [emu] Move host.architecture to CpuArchitecture
In the attached bug, KVM invocation was broken because we attempted to
compare the ARCH const to a string with a typographical error. While the
bug is fixed in the short-term by fixing the typo, we would like to
avoid such errors in the future.

This CL transitions the emu_config.host.architecture and host.os fields
to use enumerations which are populated with the more idiomatic rust
from/into methods. We then add a suite of unit tests to validate the KVM
code is acting appropriately on the new types.

Because this changes the serilized form of the emulator configuration,
we also add a custom deserilizer for the architecture field which
recognizes both serialized forms. This can be removed after a few weeks,
when we believe users are unlikely to have emulators running with the
old format.

Original-Bug: 99100
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/674057
Original-Revision: f1445794032e5285ba9e67b01695209a3ce97cf0
GitOrigin-RevId: d67585832cc5b29b2f6b99d4e7fb7b3f7835f4b9
Change-Id: I9b6865846ec3e4476c57017a2d8f3c87d5415dd2