[roll] Roll experiences [ermine] Fix for screen flicker with overlays on top of child views
The screen flickers when overlays are visible on top of child views.
Ermine draws a semi-transparent scrim that darkens the content of the
child view in order to make the overlays stand out. But for some reason
it has causes flickering.

This is being investigated at the Graphics layers. Until then we are
removing the semi-transparent scrim.

Original-Bug: 99036
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/experiences/+/674608
Original-Revision: b379dfdaa275d4b59ed250dd6214a614c7bb8eee
GitOrigin-RevId: b40a788b81c5810bbd07d796f4ac3409c14a5253
Change-Id: Ia00ba974ea3a6e58af75b8cf9f1cd85710ec32b2