[roll] Roll fuchsia [pci] Don't assert if root_ isn't allocated in dtor
OOM testing demonstrated a situation where the Bus ctor path fails to
allocate the root node. In this case we properly assert that it wasn't
allocated in the dtor. This change switches over to verifying
allocations via alloc checker instead, and no longer asserts in the case
that root_ doesn't exist when we tear down the bus.

Original-Bug: 99076
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/674357
Original-Revision: 30de446e571c79405c79145d370bf97b89520183
GitOrigin-RevId: 89db1136d2bc9bb2e62b56c51cb24842639cebab
Change-Id: Ie303e44c0bc3efdbfc125b58e259e3c086512cf6