[roll] Roll fuchsia [fdf][ramdisk] Add wait_for_device to biotime test
The biotime test occasionally flakes because the
test can run before the ramdisk is made available,
causing the ramdisk_create to fail.  This CL adds
a wait to the setup of the test to make sure that
the ramdisk is available.

Testing: this is a change to a test.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/674702
Original-Revision: 2514baeabdcdd720c8359c7cfd083e45d76d7129
GitOrigin-RevId: f7de86c3b69695fd381ce72e256867a86dee8767
Change-Id: I5870ccde9c311a460e8075fb96440388c7dd1c1b