[roll] Roll fuchsia [inspect] Add AtomicUpdate to C++ library
Add Node::AtomicUpdate(), which allows the caller to perform multiple
operations that are guaranteed to appear in the same generation.

Introduces State::{Begin,End}Transaction() to support this new
mechanism. While one or more transactions are open on a given State, all
operations will appear in the same generation -- regardless of the
thread from which they are performed. This is aligned with the Rust
library semantics.

Original-Bug: 97904
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/670768
Original-Revision: eabf4fc79b74f5f7f0c1cdc03475e16c23392bd8
GitOrigin-RevId: 9c7fc07f5b5636e78b505b45a8733fc7c368e8f4
Change-Id: I55e4cddabad60aba5736a6b674dc623d6636ea5d