[roll] Roll infra/recipes [testing_requests] Assert shard deps exist
Sometimes a change will break the build in such a way that not all of a
shard's declared deps are built. Currently, this results in an infra
failure when creating symlinks, e.g.
https://ci.chromium.org/b/8815270772206965345

However, these are really user-caused failures so they should cause step
failures instead. To accomplish this, validate that each dependency
exists before adding it to the symlink tree.

Unfortunately it's tricky to get test coverage of this path due to the
awkwardness of mocking which paths exist, but I ran the tests locally
without the `mock_add_paths` call and it produced the expected error
message.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/674964
Original-Revision: cfa085b94567e54e988306c57fda4fc89d9e5325
GitOrigin-RevId: 4528ee889fe23c5dce70a838d67df7e7a7ac26d7
Change-Id: I6b02245dce487fee65c811c6249247b65b66fd06