[roll] Roll fuchsia [blackout] use ffx to run commands on the target
Set up an isolated ffx instance to communicate with the target device
and run the target components. This removes the need to call ssh on our
own, or do target discovery.

As part of this change, we swap the meanings of some return codes,
making verification failures need a specific return code and retrying on
any other errors, making the retry logic more resilient.

This also adds an integration test for blackout that runs in CQ,
providing a backstop for the host-target-interaction functionality.

Original-Bug: 44392
Original-Bug: 96569
Test: ffx storage blackout blobfs_checkerboard
Test: ffx storage blackout integration
Test: fx test --e2e blackout-integration
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/662499
Original-Revision: 7eb72d8e67542c831c213730141f001cc575da25
GitOrigin-RevId: 4e61a4568f90509bf733598b55df25db2f4c978a
Change-Id: I1742cc8980533e8727ee7a14f976a84efc75956c