[roll] Roll fuchsia [cpp][style] Guidance for naming generated impl headers.
This has come up as a topic in the RFC for configuration accessors
which need to generate these headers. They will not go into an
SDK, and are generally "private" to the component using them.
The current guidance for generated headers assumes they are part
of the platform, and only provides guidance for `fidl/`, which
does not (currently) describe structured config libraries.

Original-Bug: 95371
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/673667
Original-Revision: 78196eccc5b290779b37ca597a082f167f458f69
GitOrigin-RevId: e0574df04b16187eb194acc47976ae4bb9202aaa
Change-Id: I771ea320c7505beada5160b62ad994de47080749