[roll] Roll fuchsia [media/vnext] add buffer lead time tracking to stream_io lib
This CL adds lead time tracking to the stream_io library. Lead time
tracking assigns a lead time to newly-allocated buffers, indicating
how far in advance of allocation the release of buffers is occurring.
Given that buffers may be allocated out-of-order with respect to when
they were freed, the arrival time of a buffer (the time at which it
is freed) is pushed into a queue. When a buffer is allocated, the
current time is compared to the arrival time at the head of the queue
to determine lead time.

Test: unit test added
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/671125
Original-Revision: 4e7e2eeea75934282739fe27138be3e73f367a56
GitOrigin-RevId: d0bfc0fe83c2d8d3c1796fa41c08c7fd7030c3cc
Change-Id: Ia3062cf8c887ed4270a05a39ae678ce17baf0757