[roll] Roll fuchsia [boringssl] Use same warning options for kernel and user builds
The kernel integration of boringssl code is so minimal that it
mostly has its own separate build logic for the little bits it
needs.  But the warning flags needed for the current compiler and
build defaults to mesh with boringssl code is something that
stays common between kernel and other compilations of boringssl
code.  So refactor the GN code slightly to share the baseline
config of warning flags between both kinds of compilation.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/672042
Original-Revision: 9ffed6e13285ae0acbedf81bb9ee0b8e023752ca
GitOrigin-RevId: 7c899f9ed9fb768196a09c6f3bfb1f649fd40e4d
Change-Id: I18beef851b96ce16322f22e11c440eb8dbc9f4e2