[roll] Roll infra/recipes [fuchsia] Only update "latest" assembly inputs ref in canary
For consistency with the SDK upload process, only update the "latest"
ref of an assembly inputs CIPD package when on the canary branch to
avoid backwards rolls that result from moving the ref from a canary
version to a version produced on an older release branch.

This logic is basically copy-pasted from the sdk recipe, except we
hardcode the full canary branch name rather than taking it from a
`ref_settings` property to avoid the complexity of plumbing it through
from the configs. This is a hack and should be cleaned up ASAP, but it's
good enough for release branches.

Original-Bug: 99452
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/674967
Original-Revision: 4a64121ace026b41b079372318711e4f78f3b710
GitOrigin-RevId: 3bb676882cec7f51e93a264a329e7b81be5326cb
Change-Id: Iac038cb81631e576cfff25d42a2d136c128881b5