[roll] Roll fuchsia [fxfs] [recovery] Overwrite Fxfs superblocks as part of FDR.
Support overwriting of Fxfs superblocks so that we can perform factory
reset of devices without zxcrypt.

This is only step one.
Following this, we need to:

 1. Move factory reset to fshost. We can then shutdown the filesystem
    before overwriting, removing a race that would prevent reformat.
 2. Implement keybag deletion under Fxfs.

We will not be in a shippable state until both these things are done but
will benefit in the interim in being able to unblock zxcrypt removal for
Fxfs native-ecrypted volumes.

Original-Bug: 98889
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/674325
Original-Revision: d5b4bc9ef3e840c8d685c8ee9a3db653517a4446
GitOrigin-RevId: 7188514655ccc1ae1021f319a1c019837da235dd
Change-Id: I937af65833d949e67d84ddc2215f3aa87ee8c6ee