[roll] Roll fuchsia [acpi] Prepare to build on the host.
This CL has no functional change. It adds a bunch of #ifdefs to the
existing ACPI code so that we can build it on the host, and a stub
implementation of acpi::Device that doesn't call into the DDK.

We also split Manager into FuchsiaManager (which basically just
implements all the async bits) and the HostManager (which stubs them
out, since they're not used on the host).

Original-Bug: 98871
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/674813
Original-Revision: 660b97c9bfb61697e8705ce35003ca919f279e63
GitOrigin-RevId: 3768f559b11241b14b6fcb69417f2791d96c18e7
Change-Id: I00f1b2ab49dbe2458b2381636caf0ed5c0d516b2