[roll] Roll fuchsia [fdf][storage] Change factory_reset_test to use ramdisk
The factory reset test uses a function in the fs_management
that will in the future assume that the file descriptor argument
refers to a block device.  Rather than alter that function
to handle both vmos and block devices, we alter this test to
have it use a block device.

Original-Bug: 91608
Test: this is a change to test code.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/674584
Original-Revision: 3b3141cb551903ef0a462d3723e2aae73f1e37e1
GitOrigin-RevId: abe3d5c17277e7dd49cdf8dff0483f92394ac820
Change-Id: I31f55ca1455e032c0db4bdf5813d8148ca379185