[roll] Roll fuchsia [codec] Fix timeout on end of stream
There seems to be a bug where std::cv_status::timeout is almost never
returned when using system_clock. In addition, we want to wait for a
certain amount of walltime to pass so we should be using steady_clock,
which fixes the bug.

Original-Fixed: 98636
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/671557
Original-Revision: 3013558848974dc43b5d394b2d895ff5c28fc880
GitOrigin-RevId: 4e3d170bc33ec294891e730f0dac06fb53b0d296
Change-Id: I76a1afadcb404d1054d17129dfb631fc32012bb2