[roll] Roll fuchsia [codec][vaapi] Implement diagnostics for VAAPI
Implemented the diagnostic data in a way that can both be used by
amlogic driver and the vaapi driver. A function was added to
CodecAdapter that allows the driver to provide the new created codec
with a reference so that it can create codec diagnostic information.
The codec is agnostic to how the diagnostic information was created
since, as of now, amlogic is a driver and vaapi is a component.

Original-Bug: 81304

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/672564
Original-Revision: 10f401460567c76c0f75111c0ae4e9e417233a62
GitOrigin-RevId: bdd9bc2fe62d7fbd375a3de8df5f661d9af72a37
Change-Id: Ie3c9a82116b8ad0e796b0ee009f281d7efb4fbc0