[roll] Roll fuchsia [ffx][test] Retain and save all suites at once
This updates the directory reporter to retain all results in memory,
and save all results when either a suite or test run completes. This
essentially removes an optimization where suites were saved and evicted
from memory as soon as they completed.
While this might regress performance somewhat, we will need to do this
anyway when we transition to a single file format in the stable test
output. This change is in a separate commit to ease reverting in case
this causes issues.

This change also removes some unit tests related to the optimization
that are now obsolete.

Testing: run CQ in the *ffx builders
Existing tests

Original-Bug: 81195

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/674748
Original-Revision: 672a3bf06e2b9dbe4d0054c6ead1a5a410f152e2
GitOrigin-RevId: 02a19daad0eb6bf91f0dda4444379c880fb9edad
Change-Id: I7a3f40dd06c1b33fe11f0bc10fec8192cfb6b000