[roll] Roll fuchsia [ffx] Add ffx profile memory.
This CL allows ffx to print a summary of how the memory is used on the
remote device.

A new FIDL api is added to allow ffx to send a socket handle to
memory_monitor.

ffx creates a socket and sends one of the two handles to memory_monitor.
memory_monitor writes all of its data to the socket.
At the same time, ffx reads all the data sent to the socket,
and then prints it out.

Test: fx test ffx_profile_memory_lib_test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/670542
Original-Revision: e368105c31992ae6d2266a6da0e7db05687e27d8
GitOrigin-RevId: c29b9c8d5fcbefb8e9ef7aec9af1513d8209c56f
Change-Id: I1bc0cbe335127dcac3ee58603a2144697008289f