[roll] Roll fuchsia [wlan][rsn][sme] Don't disconnect after key frame retries
Some APs will not complete the EAPOL exchange if the second EAPOL
message is sent too rapidly. It is better in that case to simply wait
for the handshake to restart again.

This change introduces new logic to inspect the state of the fourway
handshake in case of an overall RSN timeout. This is so SME can report
the reason the handshake could not proceed, including the case when a
bad credential is suspected.

Original-Bug: b/229401143, b/203003309
Test: Passes ACTS WlanSecurityComplianceABGTest.
      Manually tested successful connection with an AP with the noted
      behavior.
      Added new unit tests.
      Manually tested bad password behavior by saving a bad password.
      fx test //src/connectivity/wlan
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/674286
Original-Revision: e7ed2c448e336d9113e370685b859e5d155719f3
GitOrigin-RevId: 3780bcb31afdc2d359e8d5aafd98a67e9f0d43f9
Change-Id: I9c2c0456ba46606e2c96c22c44a86760f16fd95f