[roll] Roll fuchsia [virtualization] Relax header size check
We were enforcing that the disk size must be aligned to the cluster
size. This is not something required by qcow. The block device already
has the sufficient checks in place to reject any request that extends
beyond the device capacity so this check can be removed.

Original-Bug: 95529
Test: Booted debian guest (required copy-on-write patches). Fixes an
      error when booting due to triggering the removed error
      condition.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/675907
Original-Revision: dd667d153702b900cff52a9edf5ba0a04979e01f
GitOrigin-RevId: 2bcf2689eead04326dc1a1dfe73c06e616d559a3
Change-Id: I7705d8ffbde9e029d0cb115255421c27581c96b8