[roll] Roll fuchsia [netstack3] Allow tentative addresses to be deprecated
Previously tentative addresses were assumed to be not deprecated since
there was no way to know the deprecated state of a
`AddressState::Tentative` - all addresses are added in the tentative
state so when a SLAAC address with zero preferred lifetime is added, it
will enter the tentative state then transition to assigned and we lose
the deprecation information.

This allows us to increase the possibility of a host having an IPv6
address. If a non-deprecated address comes along later, it will be
more preferred than the deprecated address (once it resolves DAD).

While I am here, drop `NdpContext::deprecate_slaac_addr`.

Original-Fixed: 99569

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/675801
Original-Revision: 8af37c4dd7b6068c3f21446a95113957dcfada56
GitOrigin-RevId: e84365796c92112c426410da2231e3092ee415f9
Change-Id: I6ebbe86e9668416532abf24c3ffbd10df76021a1