[roll] Roll fuchsia [a11y] Fix flake in flatland-a11y-view-tests
This test constructs a flatland display, and attaches the a11y view and
a child view of the a11y view ("the proxy view"). The test fixture
determines that the a11y view has been inserted by waiting until the
proxy view is connected to the display. Prior to this change, the test
only polled the proxy view's parent viewport status once. In most cases,
the first status reported was CONNECTED_TO_DISPLAY, and the test passed.
However, it was possible for flatland to report DISCONNECTED_FROM_DISPLAY,
as the first status. In this case, the test would never call
ParentViewportWatcher::GetStatus() again, and would hang.

This change ensures that the test fixture continues to poll the proxy
viewport watcher until the proxy view is connected.

Test: flatland-a11y-view-test
MULTIPLY: flatland-a11y-view-test: 200
Original-Bug: 99629
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/676323
Original-Revision: 069ff608000f325fc0b04b0f68ec1c05ee595281
GitOrigin-RevId: c95dc4b3f3ae1e778b9fd1e795613076cb8fd8e1
Change-Id: I6707c9686c699b560e0b4a2a0d8992ca3979aa81