[roll] Roll fuchsia [wlan][hw-sim] Adjust beacon period in test
Almost all our hw-sim tests have a long beacon period set in the beacon
frame to counteract against slow bots. However, this is ignored in the
multiple-clients-ap test which uses the SME API instead of Policy API,
and which reads the beacon period from the ConnectRequest.

This commit sets the beacon period in the ConnectRequest to the same
amount that we set in other hw-sim tests.

Original-Fixed: fxbug.dev/99596
Test: Quick check locally that test still succeeds
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/676341
Original-Revision: 2c5e4f420665d5056a3835c377e850b4b8972a35
GitOrigin-RevId: b8782a71e008e7fd8dbd70e2d6aa9bcbf8fc2ea5
Change-Id: Ib0a87b43ba7e88b8a5abc1c93944424f2ff68b28