[roll] Roll fuchsia [check-licenses] Support explicitly adding licenses to the NOTICE file.
We currently use the output of "gn desc" to filter the full NOTICE down
to only include license files that actually contributed to the build.
This works well, but there are some situations when we want to add additional license texts to the NOTICE file, e.g. b/229106823.

This CL modifies check-licenses to take a list of "filter"s, which is a list of string paths to JSON files. The JSON file format is a simple string list, where each entry is a path to a license file or project. Check-licenses will read through these files, and include every associated project in the final NOTICE file.

Original-Bug: b/229106823

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/675057
Original-Revision: 7b2075a35bfe5a1ab8bb485b2b6725514a7d8e37
GitOrigin-RevId: 1a1f5183760e97a89883e9a1ae5231804f7555c1
Change-Id: Iacd1e8526557624f13737d4029d5ba52889cf85a