[roll] Roll fuchsia [fidl][cpp] Populate dynamic_flags
Change `fidl_init_txn_header` to accept an additional parameter for the
dynamic_flags for the call. Update callers and the generated fidl
bindings to supply this extra argument. In the case of generated
C++ bindings, populate dynamic_flags with the correct strict/flexible
flag for the methods. For the deprecated C bindings, always populate
with `FIDL_MESSAGE_HEADER_DYNAMIC_FLAGS_STRICT_METHOD`.

This change opts to modify `fidl_init_txn_header` in order to require a
conscious choice on the part of the caller on whether to spcify strict
or flexibe method because strict vs flexible methods are permanent
feature of the language, not just a migration stage.

An alternative design choice would be to leave `fidl_init_txn_header`
as-is and require users to specify `dynamic_flags` afterward if needed.
This is what is done with `at_rest_flags` (formerly just `flags`), when
specifying e.g. encoding version flags.

However, in the case of at_rest_flags, and especially the fidl version
v2 flag, the flag is a temporary migration feature, so it makes sense to
have a default value for the flag setting, and then manually override it
when needed at different stages of the migration. Since dynamic_flags is
a permanent feature  for supplying information about each call, it makes
sense to make it a standard part of initializing the transaction header.

For type-safety in C++, introduce a fidl/cpp/transaction_header.h which
complements fidl/txn_header.h and defines `fidl::InitTxnHeader` a
version of `fidl_init_txn_header` which uses a new enum class
`fidl::MessageDynamicFlags` instead of plain `uint8_t`.

Original-Bug: 88366
Test: fx test fidl_hlcpp_unit_tests fidl_cpp_unknown_interactions_tests fidl_llcpp_unknown_interactions_tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/672580
Original-Revision: 3e3d363e33df539bbaa61f11dcf67ec1706d4270
GitOrigin-RevId: b8ae3a7bf86fca26cec38da97b23e1bbd79b4142
Change-Id: Ic268b01d1321acfc3695fad81f1e8bf9b031a3e0