[roll] Roll infra/recipes [checkout] Skip patching integration if not in checkout
It's currently not possible to test a change to a smart-integration
builder's properties in integration.git CQ, because the recipe will
refuse to patch in an integration change since it's not in the checkout:
ci.chromium.org/b/8815054002445456033

However, recipe_bootstrap will have already applied the patch when
evaluating properties, so the recipe should just skip the patch.

We are already lenient about this when it comes to pure Git checkouts,
this makes the Jiri checkout path more lenient as well.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/676627
Original-Revision: e7e6bcd6760d122075555147679f0679c08f2cc6
GitOrigin-RevId: afe2b6baec2b428803f6dbd734ea422d085be8c1
Change-Id: I04c1e452ff93ecb2e136557e337484e9fe30439e