[roll] Roll fuchsia [testrunner] Create stdio file even if empty
Currently, if a test doesn't produce any stdio then testrunner will skip
writing a stdout-and-stderr.txt file, but still declare in summary.json
that such a file exists. This causes failures when the recipe tries to
read the file that doesn't exist, e.g.
ci.chromium.org/b/8815013275155244257.

For simplicity, we now always write the file even if it's empty. I also
considered not adding the file to summary.json if it was empty, but that
made testrunner's unit tests more complicated because the mock data
needed to conditionally include stdout-and-stderr.txt depending on
whether the mock output was empty.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/676783
Original-Revision: 37f7e061601ed5314d289207d02bcd13c949506e
GitOrigin-RevId: 801f4b0e8d893e80dd2b28e0d24c5123551f97ea
Change-Id: Ic25725ace75f0cccee824c2a1ffc1ccd91dd4ff5