[roll] Roll fuchsia [dart] Remove unused acts_test
This test harness isn't being run by CI/CQ currently, and won't be for
a little while.  The author is willing to convert it to another language
when they work on it again, so delete it now to reduce our dart
footprint in-tree.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/668468
Original-Revision: 6e46d0f4816d7aa65d940c5080f65ae88fd83ca0
GitOrigin-RevId: 02e5190b38bf0cb159282004279298ce16cab531
Change-Id: I1925add4c9df01f09c8e701c834ae693cc7eb388