[roll] Roll infra/recipes [btu] Fix tempdir naming scheme
Including slashes in a temporary directory prefix is not safe because
it will be interpreted as multiple levels of directories, which will
case `mkdtemp()` to fail.

For example: ci.chromium.org/b/8814994983934690897

There's no reason the directory prefix has to be based on the package
name, so instead we just use a constant prefix.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/676862
Original-Revision: b4ad7e5fc666b651723f2771cfde808d32a8e2b1
GitOrigin-RevId: 5a94a2cea87e3501faf7266258c5414f891e8034
Change-Id: I103128f195fc1674cf2c6b2ba161f0de139e515d