[roll] Roll fuchsia [kazoo] Print instructions before the diff on fail
In the kazoo golden test it would previously print the diff followed by
instructions on how to update the golden file. However, if there are
many failures the output will get truncated by the bots and the
instructions will be lost.

This prints the instructions first so they will not be lost in buildbot
output.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/676842
Original-Revision: ed749113ba1e5ed63028c7e57946a0909ee405d2
GitOrigin-RevId: faa42a599a415fbf7a2c2b02e380bbd30103e59b
Change-Id: Id27448090aeddc3523355ee0afe2eab0ebb23ee4