[roll] Roll fuchsia [codec] Replace system_clock with steady_clock
In order to wait for a set amount of time we should use steady_clock instead
of system_clock. steady_clock is guaranteed to be a monotonic clock. Currently
we were using system_clock which is only guaranteed to be monotonic iff
is_steady is true which is platform dependent.

Original-Fixed: 99758

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/675345
Original-Revision: 4aa6b5094d508d87f019df84a5629f90cc7e9a72
GitOrigin-RevId: 8d136aac9af203b633f6cc88012d9ac4ec7879b5
Change-Id: I9c21a1f6486ea21e43d25ed029ee133cebb1a589