[roll] Roll fuchsia [fxfs] Track allocations by owner_object_id.
This is a breaking change (Fxfs version 15->16).

The owner is the ObjectStore that holds a file so
this change tracks the bytes allocated to each ObjectStore.

This information will be used when deleting encrypted ObjectStore
as well as in enforcing storage quota limits.

Note that the space consumed by a volume is more than just
the bytes allocated to owner_object_id for the volume's ObjectStore.
It should also include the StoreInfo file and layer files (stored in
the Root ObjectStore).

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/675328
Original-Revision: 599baf20d53467abffaf23fdf961e357914d5686
GitOrigin-RevId: 670b6a77ca01adb8e81facf758008d55ec4b50c7
Change-Id: I6e4eddb3d6be263ce51c8da755c9b97ca7985182