[roll] Roll fuchsia [kernel][socket] Opt sockets out of tracking
Removes the socket dispatcher lock from the active list tracking,
allowing the lock to be held even if lockdep::AssertNoLocksHeld is
called. This is intended to be a temporary solution so that the
default can be made to deny holding locks while user copying.

Original-Bug: 99589, 94078

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/676125
Original-Revision: 58990db2b968e3f2492841715db41c44e5ba14e1
GitOrigin-RevId: 303eeb4bca864711033e6243593f5929b3fdc6bc
Change-Id: I7cc90703e82136b702ded9d86fa019e2813d3a6e