[roll] Roll experiences [ermine] Fix gray screen when app does not load from package server
This change fixes an issue when an app is launched, in this case,
spinning_square_view, but fx serve is not running. This results in a
gray screen and the user cannot bring the system overlays back.

The fix is to allow user to invoke system overlays even while waiting
for the application to launch. This works because the shell still has
focus. And when the app does launch, the system overlay is dismissed
when the app receives focus.

Original-Fixed: 99699
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/experiences/+/677006
Original-Revision: 1d994c619670ba83cf0d5a200d18962e58c75ccf
GitOrigin-RevId: bcce46113725ef0040278ab96c3436eeb4463ed6
Change-Id: I581ceadc6220a37e052ab58384fc68a3dce58c5d