[roll] Roll fuchsia [fxfs] Add support for volume creation/enumeration
Implements the Volumes API on Fxfs.

Also adds the volumes directory, which relies on
vfs::directory::immutable::Simple and as such inherits most directory
functionality for free.  Note that this isn't wired up into the export
root yet, that needs to be done in tandem with fs_management changes to
adjust how Fxfs is launched.

Also adjusts the naming of a few methods in the API to match the RFC,
and moves removal to be within Volumes so that we can internally rely on
the VFS' Simple but still retain control over how removal works.

Original-Bug: 99182
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/676342
Original-Revision: d66f0f399afb20ea3a1e32d52d8b86b83b6fe222
GitOrigin-RevId: c3ae56dd4056623f569af2a2ff5d2ce3b6a00d2e
Change-Id: Ibaaa473657337c2edf91ff1fb8e77be990172e81