[roll] Roll fuchsia [overnet][fasync] Cancel redundant overnet port waits
Overnet signal proxying creates a lot of OnSignals objects, each of
which calls zx_object_wait_async on creation. But if the handle in
question is never closed, and the event never happens, the port
registration sticks around even after the OnSignals is dropped and the
packet would thus not have anything to handle it. Eventually the linked
list of port registrations gets too big and the kernel kills
overnetstack.

This CL adds a Drop impl to OnSignals which cancels the port wait. This
requires holding a reference to the handle inside of OnSignals, which
requires adding a lot of explicit lifetimes in overnet to prove that the
handle isn't closed while the OnSignals exists.

Original-Fixed: 95247
Original-Fixed: 96326

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/675183
Original-Revision: 205342cc930f6a9108ae04d23db64b913f7ab795
GitOrigin-RevId: c5adc1062c5f18475012e3e47469189108841e25
Change-Id: I2d4eb5d81ac641e24e5cf91968e510f040e005eb