[roll] Roll fuchsia [settings] Move from //garnet to //src
This change moves the source code of the settings service from //garnet
to //src as //garnet is deprecated. Proxy targets are left behind in
//garnet so that existing workflows and references don't break.

Some files in the meta folder are not being migrated in this CL due to
dependencies on the actual file path, which can't be solved with proxy
targets. Those will be untangled separately.

Original-Fixed: 98248
Test: CQ, verify that developer workflows still work locally
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/676811
Original-Revision: 0911a4f17bdcb1dae0ea0a5797508914b5bb9f96
GitOrigin-RevId: 50f9a81cfc1ff3488632899bb05b1ae2dedd717d
Change-Id: I233f60ee6491b3cbd70eb39cfc6ec1b3bfb02209