[roll] Roll fuchsia [fidl] Remove NaturalBodyEncoder::GetBodyView
Before this CL, a BodyView struct was created and then the values were
copied to an OutgoingMessage. After this CL, the OutgoingMessage is
directly created.

The original goal of this CL was to improve performance, but it
doesn't have much effect. Still, this CL does simplify things.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/669550
Original-Revision: 36bf3e3af54813ac9f11c949ff754f0db9312d11
GitOrigin-RevId: ed1c69377c44c84e32c95e1d269f4a3f6e5fcbf2
Change-Id: Idbf02b177bbb080568e572bf99f248a9b208b5e4