[roll] Roll fuchsia [bt][fidl] Introduce the sys.Pairing API
Define a new protocol, `sys.Pairing` to allow system clients to set the
PairingDelegate. This protocol enables and facilitates Bluetooth
pairing.

The `Pairing` protocol provides the `SetPairingDelegate` method. This
method is identical to that in `sys.Access`. The types, invariants, and
semantics remain unchanged.

Specify versioning for the newly added protocol as well as a note about
the deprecation of the sys.Access variant.

Original-Bug: 98413
Test: fx build

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/672987
Original-Revision: 2145c54cabe4decedde928be923b32405e2ea086
GitOrigin-RevId: b065d80fa4664a488e418cf58f1ea621b787839c
Change-Id: Ia1622add4c5fad06e0afd4b34eb793c2abf7dac9