[roll] Roll fuchsia [codec][vaapi] Fix race allocating output buffers
ProcessOutput may be called before the sysmem buffers are allocated.
The code to determine the output buffer size could assert because it
depends on the output buffer constraints.

If we move that calculation code after
output_buffer_pool_.AllocateBuffers then it's guaranteed that the
buffers are allocated by that point.

Original-Fixed: 99841
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/677085
Original-Revision: 974efd9f18939210c60e62484d8c92a7fcae7777
GitOrigin-RevId: f4007a9a134d0af4dcbfb5aede15509cb41e0338
Change-Id: I836a4a99c9ee0be346d899a95a2cbae265823fc0