[roll] Roll fuchsia [fxt] Added Writer and Serializer Classes
Added interfaces to allow writing of fxt to a generic buffer. To start,
this just adds a single FXT method to the serializer to first get
feedback on the interface. More will be added in following patches after
the interface is settled on.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/670187
Original-Revision: d1cb870a7a3552095b6664c603a8d4ee2b44b7cc
GitOrigin-RevId: 0853a6239f86a01c70e16aa000c23422baf00295
Change-Id: I14a2f03f84879260b655b3efb3334d1aa5fbfd1f