[roll] Roll fuchsia [bt][gap] Coalesce peer name from many sources
Replace SetName() with RegisterName(). Whenever a peer name is read,
register the name along with its source location. Store the name which
originates from the highest priority source in the peer as the canonical
device name.

In AdvertisingData, replace LocalName with a struct that bundles the
local name with a boolean that is true if the name is complete or false
if it is shortened.

Original-Bug: 90213
Test: fx test bt-host-common-tests bt-host-gap-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/657444
Original-Revision: f3cce7cb3c65748ed48be87b8049a14b81583ca7
GitOrigin-RevId: 229617cfae33a2bac08fe0e668e2a459d4ee7573
Change-Id: Ic8645ad8b05f90f2f0bcf439f915c3f314079915