[roll] Roll fuchsia [Brightness manager] Auto-brightness in Brightness Manager.
This CL add a small, auto-brightness component to be used in basic
systems. It dims and brightens the screen and the LED(s) as
the light on the light sensor changes.

It is very simple and completely self-contained.
There are no controls or options for it.

It is a separate target (auto-brightness) in the build file and
it contains a shard that can be added to existing configurations
as required.

Starts automatically if included in a build.

Common files between Brightness manager and Auto-brightness have
been moved to a library.

Test: New tests added, all tests pass
Original-Bug: b/226527434
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/673306
Original-Revision: 3a152eef0860b19e7b94718fb6cc73f0c3c7e820
GitOrigin-RevId: 725734310ef552899784d41bd2d673b2609ae12c
Change-Id: I4d41ca19043876dc0d6f2e9cd4f1d7dd53fab6d2