[roll] Roll fuchsia [mediaplayer] remove reference to buffer_internal.h
mediaplayer used to reference ffmpeg's buffer_internal.h in order to
access a field in an AVBuffer. This is a problem,
because buffer_internal.h reference <stdatomic>, which is not
compatible with our C++ toolchain going forward.

Fortunately, there is a public method to access the field in
question, so the reference to buffer_internal.h may be removed.

Test: no behavior change

Original-Fixed: 99837
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/677802
Original-Revision: 9899bf712512eea76205310388b13ed1f81f76bb
GitOrigin-RevId: c5dd01d5d062379c7a4aadb85de22835bffe6382
Change-Id: I619b225a138e49014b126527330ca21c9b36a145