[roll] Roll fuchsia [kernel][socket] Reset pkt_len_ in MBuf
The pkt_len_ field is used by the Read code to determine if a given
MBuf is part of the datagram chain or not. Since pkt_len_ is not reset
when a MBuf is recycled, reading and writing different sized datagrams
could cause a datagram head packet (with a pkt_len_ set) to get reused
as a datagram body packet.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/677864
Original-Revision: 9e31959fcc1b981565a645e8745032b2cdd7defd
GitOrigin-RevId: 1049524571b22229d2f7525d938653297a38f41f
Change-Id: I0e3382329ab6f514da5f00584cfd21f3d80c87e2