[roll] Roll fuchsia [assembly] Add fuchia legacy input bundle archive to uploads
Add //build/images/fuchsia:fuchsia_legacy_input_bundle to the set of bundles
that are documented in the build_api.

Threads a parameter on assemble_system() through to the creation of the
legacy bundle, and then only sets that flag on the main fuchsia assembly.

Also changes legacy.gni to create the package and archive using deps
instead of just creating the targets (but not depending on them).

Fixes: 100066
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/678296
Original-Revision: dbc2b6cd6b3152776f6de86ebc674e8ccdd54683
GitOrigin-RevId: a4f4272db5479a477d4fbfe54df885cc10b89aa3
Change-Id: Ib8152b4d298c53748f4761cfe5924840201bf74c