[roll] Roll fuchsia [tools] Fix VM RAM allocation in //tools/emulator
Remove support for VMs with <1MB of RAM. QEMU 6.2.0 (which we now use)
doesn't support values that low.

At some point, QEMU began to expect values provided for the
`-m` command-line switch to be an integral number of megabytes --
not bytes. Our code assumed the latter, which meant that the default
value of `8K` used in //tools/emulator got parsed to `8192` and
handed off to QEMU...which happily started a VM with 8GB of RAM.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/678262
Original-Revision: bc62af0bb3b593e616e2afc2c647a3bdea47f925
GitOrigin-RevId: c46c9c0f3f8c584dc2cd38a96588ce9eb06eac53
Change-Id: I1afc5cd9739f47a1599e9f7f1474420693454e7c