[roll] Roll fuchsia [physboot] Increase mexec hand-off scratch space
This space is currently limited to a page (seemed large enough at the
time). We are hitting this limit now, however: a particularly fragmented
EFI-level view of the physical address space is lending itself to a
MEM_CONFIG ZBI item that is just about a page in size itself, which we
then fail to account for in the scratch space.

A more robust fix is coming, but this is a good enough to address the
associated bug for now.

Original-Bug: 99875
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/678225
Original-Revision: c1a6b68074b9423d09c62cf7a25fbfb7ba8cc883
GitOrigin-RevId: 9d0e49a7af2faebd44f94387579efb750cc6e582
Change-Id: I2f7ba4d369b658c9eea618258afb07926a58ceaa