[roll] Roll fuchsia [kernel][mmu] Expand consistency check for arm64 PTs
The purpose of this change is to expand our consistency checking to
help isolate a seemingly map-after-destroy issue we're seeing in the
field.

Add a new method, DisableUpdates, to ArchVmAspaceInterface that puts
the arch aspace into read-only mode and asserts that it contains no
mappings.

Call DisableUpdates as part of VmAspace::Destroy.

Update ArmArchVmAspace's methods that modify PTs to assert that
updates are enabled.

No functional change to x86.  DisableUpdates is a no-op on x86.

Original-Bug: 79118
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/678097
Original-Revision: 70b537214fef59380656c0e4269577d9f1426c2a
GitOrigin-RevId: 99c9b5cf3c0176774a7170d36073b79e093c999c
Change-Id: Ic54985362ba9569bbe7e38740496fbe2ac63da2b