[roll] Roll fuchsia [kernel] Pass process to HandleTable methods that enforce policy
Most HandleTable methods enforce job policy by terminating a process
that uses a bad handle value.  Currently, a HandleTable is owned by
exactly one process so it makes sense that HandleTable keeps a
reference to its owning process and simply calls the
EnforceBasicPolicy method on it when necessary.

In the future a HandleTable may be shared among several processes.  In
a world where a HandleTable may be shared, there is no single onwing
process.  We'll need a way for HandleTable to call EnforceBasicPolicy
on the proper target.  An obvious solution for HandleTable to use the
"ambient" process via static method ProcessDispatcher::GetCurrent.
This approach is somewhat sublte and could lead to surprises down the
line (what happens when a pure kernel thread, with no user mode
component, wants to call one of these methods that assume a current
process?).

This CL provides a different solution.  Instead of using a field
pointing to the owning process or using the ambient process,
HandleTable now requires callers to pass in a ProcessDispatcher& when
they call policy enforcing methods.

Why pass by reference?  Pass by reference to drive home the point that
it's non-optional and to make it clear that the callee is just getting
to "borrow" for the duration of the call.

Remove the unused HandleTable::IsHandleValid method.

Original-Bug: 98922
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/677650
Original-Revision: 60db53112ae53520975b4afe48b4a63503486afb
GitOrigin-RevId: 2b88ec247806c03d8a895a5466a6fdb401d6e9dd
Change-Id: I61aaa4e214206cbb92a96921f8aa46164a6be594