[roll] Roll fuchsia [lsan][fonts] Suppress LSAN on font_provider
font_provider reports a LSAN issue with a long-lived thread
with no good way to fix the issue immediately.

Suppressing the issue because it should not be relevant in
production, and only until we find a way to fix.

Original-Bug: 99760
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/678506
Original-Revision: aafdffd74afd6bc0ed0cacfb1d79ff95bcf28e16
GitOrigin-RevId: 501aad2fbc5411a60bcaabdd0db2e96da53429d2
Change-Id: I0830c1dc46e77f7ccf0e0862074eb6ad146e4f9b