[roll] Roll fuchsia [ffx] Add `target add` auto-wait.
This adds support for auto-wait. Auto-wait is off by default (to be
enabled later on).

The implementation for `wait` is functionally the
same as running `ffx target wait` immediately after `ffx target add`
without a wait invocation, though it happens internally.

The behavior of `ffx target add` is also changed so that if waiting, and
the added target eventually fails to connect, the target will be removed
and the client will receive the error caused during the connection
process. The target will also be removed if the client disconnects in
the middle of adding the target.

Original-Bug: 89546

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/671732
Original-Revision: e14e3505f312d4ea69f23408d1c0d98b5b667f75
GitOrigin-RevId: b53ef406d2e848a57f152a2d47b8397a07ea0240
Change-Id: I91b016b8c65cb98fd20d2bda7df5d16ba5585ed0