[roll] Roll fuchsia [memory] Use the noexcept version of std::filesystem::exists
The old code would throw an exception if the monitored file
does not exist, causing the memory monitor to crash.

This does not seem necessary, so I opted to use the
overload that does not throw an exception, allowing the
program to print an informative message.

With this realization, it seems that this crash might
be a red herring to the reported bug.  Except if
memory monitor crashing pulls with it the web engine
as well, which IIRC requires it as a service.

Original-Bug: 99414
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/677704
Original-Revision: f61bb58aed2e7486b549e3ec7929edbfafe41a76
GitOrigin-RevId: ed54a29c94ccb1d5f4b218cf173f397ca73c2486
Change-Id: If4537e398a6d606145eed7e78a1f78a4ccadd5f8