[roll] Roll fuchsia [debugger] Get relative paths from line tables
This is the intended behavior since we always want to get the relative
path against the compilation directory.  It was not a problem because
for in-tree binaries, the compilation directory is always "." thus the
absolute path and the relative path is the same.  However, in Bazel SDK
the compilation directory is set to "/proc/self/cwd", which we have to
ignore and replace with the real build_dir from the symbol index.

Original-Bug: 99751
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/678509
Original-Revision: 42c4754ad62c5b840594fff2c2489b93559e26e2
GitOrigin-RevId: 5b894282c71acfbec749f29f9d32353f7f147874
Change-Id: Ib75dca04059a321796b2ade86ebf2c332e7fa2e8