[roll] Roll fuchsia [fx] Remove dependency on python
fx self-tests currently rely on python if 'realpath' is not available,
for example on MacOS. However, python is not always available on our
trybots and should not be relied upon. This CL changes the
implementation to use a shell-based fallback that walks up the parents
of the given path until it finds a valid parent, then uses cd/pwd on it,
and concatenates with the rest of the given path.

Cq-Include-Trybots: luci.fuchsia.try:core.x64-host_test_only-mac
Original-Bug: 99598
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/678645
Original-Revision: 2a8139fd2f7c529e10c9f1182dcddfa870e8e74d
GitOrigin-RevId: dfdf8f4d70125cdf4445abd73bed88f666c831d3
Change-Id: Ib75c36cb68ed2953b1fdc16e949e07ab3c8d20e3