[roll] Roll fuchsia [zircon][libc] Update __sanitizer_memory_snapshot callback documentation
We hit deadlock when we take a snapshot where a callback waits on a
futex that would be woken by a suspended thread. This is possible if a
newly created thread is suspended in the middle of a call to printf and
the main thread attempts to call printf in a callback. Because the lock
for stdout is held by the suspended thread, the main thread will never
be able to print until it gets the lock back.

This type of situation is more an issue with how the user uses
__sanitizer_memory_snapshot. The callbacks should ideally just iterate
and read over memory regions provided to the callbacks and nothing more.
This updates the documentation so users are more aware of this.

Original-Fixed: 98237
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/671555
Original-Revision: 68e0550cb65497084d1230ae5197de12f8a6fd74
GitOrigin-RevId: aa8be5c46bc4505d552851463e145bc1186a0ebc
Change-Id: I668eb7027258715bf50941ecc4c811d9c6c2f0dc