[roll] Roll fuchsia [zircon][sanitizer] Ensure memory snaposhots cover the pthread start_arg
We can end up in a situation where an lsan scan happens at the end of a
main thread, but the before the new thread is suspended and hasn't even
started. If so, then the argument provided to pthread_create won't be
scanned because the thread register for the new thread wasn't setup yet.
To fix this, we can iterate over the start_arg member in pthread structs
in the global thread list.

Original-Bug: 66819
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/666554
Original-Revision: 126797f77045ae5906f6dbf367b7c3c4f582eaae
GitOrigin-RevId: cb985b3aa59de2f96620a7ec40cee0d5a0def880
Change-Id: I49d8351bc5c36be4286251087085becf65e72e96