[roll] Roll fuchsia [security] Allow ambient executability for workstation elements.
This unblocks a runnerless developer workflow for Flutter on Fuchsia
by giving JIT Flutter apps a place to run.

We want a runnerless workflow because Dart has no ABI compatibility.
With a runner-based workflow this forces the Flutter runner and Flutter
components to be created using the same Dart SDK which is inconvenient
for OOT workflows.

The non-vendored component_manager_policy.json5 file is only used _eng builds so it should be fine for the allowlist entry to live in this file,
but we should investigate ways of defining our allowlist that explicitly
declare what configuration they're allowed to be used for.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/678962
Original-Revision: 0acb01116c5bbb70eb8828480853f81e0785fa0d
GitOrigin-RevId: d4f83aa7ab434f35f449e7ce6e9418fe01e68724
Change-Id: Ie71a0d86e270075b864d1bffe255a3150de4d32b