[roll] Roll fuchsia [ffx][test] Move directory schema specifics into directory library
Moves all details about the structure of the directory output format
into the directory library. Before this change, the library contained
only the schemas for the JSON files contained in the output format, and
clients (run_test_suite and its integration tests) were expected to
understand the structure of the directory.
The transition to the stable output format changes the number of JSON
files present in the output format. Moving these details to a single
library simplifies this transition somewhat. It also gives us a
mechanism to specify which version of the format to use while we
migrate clients.

Original-Bug: 81195
Testing: fx test //src/sys/run-test-suite

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/647349
Original-Revision: 984150d55170dc9a7724b869637f4f618fc5bb11
GitOrigin-RevId: 83ef019606d77e972218e81c6609a5a83df3142d
Change-Id: Ib707fd21a6fdf5e7d187d4465c2797002c3ac45d