[roll] Roll llvm-project/libc [libc] add printf converter
This adds the main pieces of the last piece of printf, the converter.
This takes the completed format section from the parser and then
converts it to a string for the writer, which is why it was the last
piece to be written. So far it supports chars and strings, but more
pieces are coming. Additionally, it supports replacing all of the
conversion functions with user supplied versions at compile time to
allow for additional functionality.

Reviewed By: sivachandra

Differential Revision: https://reviews.llvm.org/D125327

GitOrigin-RevId: 0f1a11578dc224f8f3c8f2feef16c4cf9666d832
Original-Revision: a6ad433b0e59b16ada194814f32fa688bb12901d
Change-Id: I0ea2c1677383193678ee57525ee43634ad7dc745