[roll] Roll fuchsia [mdns] allow publication to outlive publishers
This CL fixes an issue in which A publication was being terminated
when the publisher that created it was closed. This required the
client to keep the ServiceInstancePublisher channel open to keep
publishers created with the PublishServiceInstance method from being
destroyed. Lifetime of publishers is more logically tied to the
lifetime of the ServiceInstancePublicationResponder channel used
by the publisher to talk to the client.

This CL gives publishers their own lifetime independent of their
creators. Lifetime is tied to the channel, so the client can stop
publication when desired.

Original-Fixed: 99755

Test: new unit test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/678223
Original-Revision: b04550f16f6d7bf8488df6d1a53190b4a94ab1f7
GitOrigin-RevId: 90f77f012d4842c88cdef9466cb839f02797aa5b
Change-Id: Ia7ddcbbd28674dcf9092bd16568d6a481ab91214