[roll] Roll fuchsia [brcmfmac] Fix tail length for SDIO frames
SDIO TX frames did not specify any tail length which caused a debug
assert to trigger when adding tail space for alignment. In practice this
is harmless because the TX buffers are guaranteed to be larger. But for
the sake of correctness and potential future changes it's better to
allocate this space.

Original-Bug: 100248
Test: Manual, sim_device_test brcmfmac_sim_test brcmfmac_test
      wlan-drivers-components-tests brcmfmac_sdio_test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/679122
Original-Revision: 67d9d4a70508297692003268c73ce2918a0e328f
GitOrigin-RevId: c322b630d3597dd511e1908b69abe3b5218d22fe
Change-Id: Ied7e36b11270ba1693ae124292da7babe7abb7a6