[roll] Roll fuchsia [virtualization] Add FIDL API for device orchestrated connections
(Note: This is an alternative to fxr/678164).

This Connect2 API would be hosted by the device, and remove the need
for a host_vsock binary. Clients would call Listen to allow for guest
initiated connections, and Connect to initiate connections to a guest.

See go/out-of-process-vsock-device for discussion.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/678662
Original-Revision: 43040ca3a47f948ed53136296ae2acd321539e79
GitOrigin-RevId: e111a78ed57d65831a0595a04396788fdc4fa21f
Change-Id: I1330a603d49bfbac9a9f420992c6d483ddf06469