[roll] Roll infra/recipes [lsc] Make lsc suggester comment unresolved (for me).
The goal is to prevent SDK and API changes from being
submitted without an `fx lsc presubmit` run. Currently
a message is posted on each CL that requires an API+1
review but this message is not an unresolved comment so it
frequently gets ignored.

Eventually we should have a proper solution that automates
`fx lsc presubmit` runs on CLs and makes the petal builders
blocking. In the meantime this is a stopgap to increase the
number of changes that `lsc presubmit` before submission.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/678682
Original-Revision: 27a2bf7f0a8fc026c1813b7ee47604841d135d4e
GitOrigin-RevId: 9e9e92f52ae1313f254dc5f23d5f5b9e7938923a
Change-Id: I505766df86cca388572bcb7d27996ee4be439ca0