[roll] Roll fuchsia [bt][gap] Do not print invalid UTF8 peer names
When a peer name is registered that is not valid UTF8, do not log the
name (just log a warning). The name may contain PII, and the way the
name is currently printed uses sstream, which is a huge dependency that
we want to avoid.

Test: fx test bt-host-gap-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/679447
Original-Revision: 69c8cb63f5ba398c234402ae548a9271b6ada5bc
GitOrigin-RevId: 3a9f66c2acc794e15603d63b45498994f07d008a
Change-Id: I45aed1b44669873e4c2ac0d359ec41d6434f5190