[roll] Roll fuchsia [zbitl] Support EFI File Protocol
This provides a zbitl::StorageTraits specialization for using
an EFI File Protocol handle as ZBI storage.  It also adds some
mock/fake support for the protocol in the efi/testing library
that's used to enable the full zbitl tests on the new backend.

As zbitl is already compiled in the kernel.phys32 environment,
this needed to clean up the EFI headers slightly to be fully
compatible with 32-bit environments (which should actually
work for 32-bit UEFI if that's ever needed).

Original-Bug: 98798
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/677302
Original-Revision: 155f81c98b983797111470f3bbf5fc5576364f01
GitOrigin-RevId: 9e1478591479e4ccc6d196a0511fba88b62c8980
Change-Id: I91e801d496b2d380536465e99774954438418404