[roll] Roll fuchsia [brcmfmac] Fix caching of SDIO block size
The block size for SDIO is cached for performance reasons since it's
accessed for every transfer that needs to be aligned to block size.
Unfortunately caching happened before the block size was actually set.
This lead to the cached value having the factory default block size. In
the past we used the factory default value so it was not a problem. As
part of the network device work the block size was reduced and the
mismatch between cached value and actual block size became apparent.

Fortunately this hasn't led to any issues since the actual block size
was lower, all that happened was that we aligned to a stricter value
than necessary.

Original-Bug: 100319
Test: manual testing, fx test brcmfmac_test brcmfmac_sdio_test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/679450
Original-Revision: 5fe4f0fc792f0e31500ffc96eb201548f51d5474
GitOrigin-RevId: 9f22682a51a19eadcb33d45439390dadd015f9ba
Change-Id: Iea56d7ca141c5a4b72aa9860965aebd0fc970b36