[roll] Roll fuchsia [kernel][vm] Fix ZeroRange for pager-backed VMOs
A ZeroRange (ZX_VMO_OP_ZERO) on a pager-backed VMO emulates a VMO write
with zeros. Previously this would decommit pages in the VMO, regardless
of the VMO type, which is incorrect for pager-backed VMOs. An empty page
slot in a pager-backed VMO indicates true absence of content, not zero
content as is the case for anonymous VMOs.

If the ZeroRange applies to the range beyond supply_zero_offset, we can
efficiently zero pages without committing any new pages as the range
starting at supply_zero_offset is implicitly zero. This is an
optimization targeting zero'ing the tail end of the VMO.

Test: core-pager-writeback
Original-Bug: 63989
Run-All-Tests: true
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/677101
Original-Revision: 170678df6fe4d2d6b6ef505580633caf0965736a
GitOrigin-RevId: 81bd26412f3d9b813f0be2c3609879b06f21233b
Change-Id: I3d5f0e2d4879299bdb50e60027458ad0474723fa