[roll] Roll infra/recipes [tree_status] Check for mid-air collisions
The first thing the tree-closer does it check the current tree status,
so it can exit early if the tree is in a certain state. If it doesn't
exit early, it spends 1-2 minutes making Buildbucket requests to
determine whether the console is healthy, before potentially updating
the tree status. During those 1-2 minutes a human might have changed the
tree status, in which case the tree-closer will clobber that manual
update. This has happened twice already since automatic reopening was
implemented.

This should never happen, so check if the tree status has changed
immediately before updating it, based on the unique key for each tree
status. There's still a window for collisions, but it's now only 1-2
seconds as opposed to 1-2 minutes.

The tree status app actually supports detecting these sorts of
collisions, but it only does collision detection for changes made via
the web UI, not via the bot-specific API endpoint. So this is a hacky
workaround for that limitation. (It probably is feasible to change the
tree status app's code, but it's a pain to update the tree status app
and would require coordinating with other tree status users to make sure
they don't depend on the current behavior.)

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/680145
Original-Revision: 1f7387731b3f825ba7aab7020fbd985f292cb3d9
GitOrigin-RevId: ed57c4315b7790845ffbada8c634cb65386dc4a4
Change-Id: I5c3050b0f6fa614b8313976ead93f33baf6f41d0