[roll] Roll fuchsia [build][rust] stop requiring output_name for rustc_binaries
This is Step 4 of the transition plan from https://fxbug.dev/64100

This makes it so that output_name defaults to the target name without
any `-` -> `_` remapping.

By now, we've been forcing output_name to be set to disambiguate between
the old behavior and the new behavior for a month, which is longer than
the week I wanted to wait (oops). So this change should be a no-op for
everyone who has updated their pending CLs sometime in the last month :)

Original-Bug: 64100
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/680140
Original-Revision: d214d175c36023cd199e8a66233e483d370f02fc
GitOrigin-RevId: f84a3ef8317410dc48920ca1b4f0db73ccbec0d5
Change-Id: I984bae5df1179f0c72a0cd1d306db20855bce6a2