[roll] Roll fuchsia [virtualization] Add Rust vsock device skeleton
This is the rough outline of the Rust vsock device (which is getting
large enough to warrant a review).

This is nearly entirely non-working, but is able to serve the relevant
FIDL protocols and accept connections.

Note that we're planning on deprecating the GuestVsockEndpoint protocol
once everything migrates to CFv2, and we'll serve GuestVsockAcceptor
directly and connect to HostVsockConnector.

Original-Bug: 97355

Tested:
fx test virtio_vsock_unittests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/668189
Original-Revision: 873b3d5f9af9e1bd6f3f3de3677197438bc1a622
GitOrigin-RevId: c21e61594c667a0e4241f10938f1206ccb057d38
Change-Id: I127398584ff6235a195bc1100ce6b273d9cb4d03