[roll] Roll fuchsia [compat] Only remove device when requested
Only remove the compat device if the driver owning the
device requested it to be removed.

The NodeController could have been unbound because
the child driver crashed (or was stopped). But in these
cases the DFv1 driver doesn't expect it's device to be
removed.

The DFv1 behavior would be to try and rebind the
child device up to 3 times, but for now we will do nothing.

Original-Bug: 100470
Original-Bug: 99885
Original-Fixed: 98819

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/679005
Original-Revision: df2c48e98afd0d49b94f775423433cd8bf24d957
GitOrigin-RevId: b8b1e5384e8bca6ff9f6cba77c3ed99fd7a0c02c
Change-Id: I3f539a012332307d16e2c36b23de2c00623912da