[roll] Roll fuchsia [assembly] Add create-flash-manifest operation
This operation currently reads the partitions config and exits, but will
soon contain the logic for constructing the flash manifest. I am running
the operation as part of the build to ensure that it continues working,
and will eventually replace the GN code for generating a flash manifest.

Original-Bug: 97962
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/678228
Original-Revision: c9701944e69db2196eef5ec261dcc317e4437880
GitOrigin-RevId: 90cb13e4e36322014dfb27ba2ff1947b041b3619
Change-Id: Ib45c67cb3f87cf559058854d20cd91f7fc968a66