[roll] Roll experiences [e2e] Adds some more error info into e2e tests
Workstation tests seem to be on the flakier side these days, but
the failure modes are somewhat opaque.  Adding these log messages
helped me pinpoint a number of interesting failure modes, as can
be seen in the attaached bug.

With this, I'd like to make the changes permanent, in the attempt
to reveal more of the failure modes.

Original-Bug: 100441
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/experiences/+/680204
Original-Revision: 0cf170e7afda2e132f6cf158fb2c7f71d262135b
GitOrigin-RevId: 621dc4819e265d131d7680e19dd39f24428c068a
Change-Id: I52b3f6e12973f8a659fdec5153091ec2d97a3600