[roll] Roll fuchsia [testrunner] Use a temp out dir for host tests.
Sometimes a host test calling ffx doesn't properly shut the daemon down
and it might still be writing to the daemon log when we try to upload
the outputs to CAS with the swarming task outputs. This causes the bot
to fail as described in the attached bug. In this case, we should use a
temp out dir when running the test and move the contents to the actual
designated out dir once the test completes.

Original-Bug: 99799
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/680306
Original-Revision: bcd911addaf84a5572526e79c0540640f417562f
GitOrigin-RevId: 539e615210d388f995b9396e6fd068aedeecb484
Change-Id: Ife74c0c4ea9ca2dcaf52167d15d8c6466ac44c73