[roll] Roll infra/recipes [build] Run all tests on recipe+integration changes
We shouldn't skip affected tests (or shard them differently) on
integration.git or recipe changes, since the affected tests analysis
will generally not be correct on those changes.

For recipes, this is because recipes are not part of the build system
and can affect the behavior of any part of a CI/CQ build. For
integration, this is because affected test analysis doesn't look at the
contents of Jiri project or package rolls.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/679361
Original-Revision: cd25399af28a943e0c3ca69b9c2796d49ebe7659
GitOrigin-RevId: 1f7e71e1157eae63d823b930df8b3d74b3a323b1
Change-Id: I81f9b2353db22b10fe6bb64662a1d7cc05e41797