[roll] Roll fuchsia [nvme] Respect CAP.MQES for I/O queues.
The spec defines CAP.MQES as the maximum number of entries allowed in an
I/O queue. The NVME driver was written assuming that MQES would always
be greater than the number of queue entries that fit in one page.

This worked fine, until it didn't. This CL adds the appropriate logic to
restrict the size of I/O queues to either 1 page OR the value in MQES.

Original-Bug: 100252
Test: `lsblk` on affected system now works.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/679854
Original-Revision: 73bf7bd8fea9dbb63661c5f999568ae2f4f0e4e7
GitOrigin-RevId: b40e450f72324b383e83c046074fa0b525124056
Change-Id: I55ebdd4dc79ac7db93f266125ac0ba83a84ae332