[roll] Roll fuchsia [kernel][scheduler] Fix Scheduler::RescheduleMask
Fix an issue in Scheduler::Reschedule mask which cause a local
preemption to fail to be recorded as pending when a thread's
preemption state's PreemptIsEnabled was false.  In some situations,
this could delay preemption on the local CPU incorrectly, in
particular, when a kernel thread dropped a contested Mutex.

Original-Fixed: 100545
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/680197
Original-Revision: 4853b2e82a21bf40ea5f03c32ff3291d0ce6effb
GitOrigin-RevId: 394c289938aef7dbca71e87c5db4797b1789b612
Change-Id: I5e9770be3f602de1be30ed423af60bfed4c3c175