[roll] Roll fuchsia [credential_manager] Add CredentialReset Protocol.
This changeset introduces a new Reset method that simply calls
ResetTree and deletes all of the on-disk state. This is part of
a chain of CLs to land PinWeaver state being cleared on FDR.
Since Reset is only used in FDR and FDR doesn't need access to
mutate credentials it made sense to split this out into its
own distinct protocol that could be routed to FDR instead of
all the sensitive credential protocols.

Original-Bug: 87687
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/678647
Original-Revision: 268d033904ace86186c108702bc2b14a83a57db9
GitOrigin-RevId: d22c0b8a6d1afff2ad2afcf5dbb7c68308720d41
Change-Id: I72d9e0b58f19b5a9bc41aaffaf79dbbcb6c11987