[roll] Roll fuchsia [fxt] Added Instant and Counter Events
This adds support for Instant and Counter Events. This also implements
serialization for Thread/String refs and Arguments as they are required
for events.

This events are implemented as described by the fxt documentation [1]

This is a reimplementation of the apis
ulib/trace-engine/context_api.cc[2] that are buffer and writer
independent.

[1] https://fuchsia.dev/fuchsia-src/reference/tracing/trace-format#event-record
[2] https://cs.opensource.google/fuchsia/fuchsia/+/master:zircon/system/ulib/trace-engine/context_api.cc;l=848;drc=0b95ec1c99b1378175e59f5a4f7ef79f6d484b0a

Original-Bug: 98177
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/675286
Original-Revision: bb662198b333165cb5905c3f6fc60c7ce3b21298
GitOrigin-RevId: dbcc2b04cdff431ea514a25ee4a504d6cb8d7cb4
Change-Id: I5176c2979742b58ca7b864af02c4690b9cce0cab