[roll] Roll fuchsia [fdf] Remove RemoteControl proxy from driver_tools
Remove the usage of RemoteControl in the driver_tools library. Instead,
functions are passed as arguments that provide proxies to FIDL services
only when needed. This is to avoid needing a connection to RemoteControl
in order to execute subcommands of the driver command. This enables the
driver command to have subcommands that do not need a connection to
RemoteControl.

Original-Bug: 1000236
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/679351
Original-Revision: 82ce40f543731cfe645222f34903347a634de240
GitOrigin-RevId: 6ee28ff37f46a4e6a53b404dcc728c8d3063536c
Change-Id: I757ea64ba041f2e6092fa19ec05c8c59a6444417