[roll] Roll infra/recipes [build] Inline postprocess_build function
It was separated out of `run_build_steps` in https://fxrev.dev/470554 so
it could be wrapped with a try/except. But the try/except has since been
deleted, so we can inline it to get rid of some plumbing.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/680551
Original-Revision: 21843ed8f445e28c2df4d4c7edc581d4846ec6d6
GitOrigin-RevId: 9890f74d826dc1e768d13e81b78bdb48ef79cee1
Change-Id: I4408a43aadbc2c9ac1b7c7c09dd4f8b49b224bba