[roll] Roll fuchsia [FIDL] Create ServiceReconnector/ServiceHubConnector
ServiceReconnector can be used to easily stay connected to a FIDL
service and operate on it with the assumption that the service is always
connected.

ServiceHubConnector can be used to easily support the service hub fidl
pattern. It handles reconnecting to the service hub and the service, as
well as retrying failed calls upon request.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/653021
Original-Revision: 7cbc4aff7548f600d8321ccb3557c37da33f5450
GitOrigin-RevId: e2f7a5ef888152e2c83f15e69782981115835203
Change-Id: I475098c8300ae3a729b60dfb1184158ee1034c3f