[roll] Roll fuchsia [fidlc] Treat absent libraries as empty
This CL allows fidlc to compile a library before its addition or after
its removal, producing the same IR as an empty library. Previously it
was an error (or a warning for a deprecated library).

Now, the library's @available annotation has no impact aside from
inheritance. In practice if a petal tries to use a library before its
addition or after its removal, it will still most likely result in a
build error, assuming code tries to use something from the bindings.

The motivation for this is to make it easier to generate API summaries
for all libraries in the SDK at a particular version without needing to
first figure out what libraries are allowed to be compiled.

Test: fx test fidl-compiler
Original-Fixed: 100604
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/680724
Original-Revision: e84552394ee914624ddda444f68fd5cb8dd61b2d
GitOrigin-RevId: decf1aa2126ea6be1c9dbc0c445295cd0adff44b
Change-Id: I0d7b0391bfd9f04fc1e198e055d131675b8098db