[roll] Roll fuchsia [fdf] Split out driver_index.fidl from fdf FIDL
The fdf FIDL is part of the SDK. The driver index FIDL is only used
internally and therefore not needed in the SDK. This change splits it
out into its own FIDL library and refactors all usages of it.

Original-Bug: 97984
Tests: Existing unit and integration tests work. New tests are
not needed as there are no changes in behavior.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/680948
Original-Revision: 222f3c211ed1df4fbd8011339472a5cde0581695
GitOrigin-RevId: 6722509f3c13cec3fc3e258493a3c251bdc48442
Change-Id: I9071075b1398b67a5ba4c7c8d7ae5c2eb86fb76b