[roll] Roll fuchsia [ffx component] Add `ffx component reload` command
Enables component reloads that update the package and manifest, while
preserving resources. The result is faster component development by
allowing fast and complete code and topology updates.

The `ffx component reload` first stops the component, then unresolves
it, returning it to the Discovered state, then starts it. The effect
of unresolving is that the next resolve will refresh both the package
and the manifest so that the component is fully updated.

In contrast to stop/start, the manifest is updated.
In contrast to stop/destroy/start, resources such as storage are
preserved.

Adds an UnresolveAction to implement the unresolve as well as an
UnresolveFailed error and Unresolved event.

Tested/Demo:
• Modify //examples/fidl/echo-realm/meta/echo_realm.cml to remove
references to echo_server. The child will not be added.
• $ fx build, $ fx serve, $ fx emu -N --headless
• $ ffx component run fuchsia-pkg://fuchsia.com/echo-rust-client-test#meta/echo_realm.cm
• $ ffx component start /core/ffx-laboratory:echo_realm/echo_client
Confirm error that the client cannot reach the server.
• Restore the echo_server child in //examples/fidl/echo-realm/meta/echo_realm.cml
• $ fx build
  [ Before this CL is submitted:
    • $ ffx component stop /core/ffx-laboratory:echo_realm
    • $ ffx component start /core/ffx-laboratory:echo_realm
    • $ ffx component start /core/ffx-laboratory:echo_realm/echo_client
    Confirm error that the client cannot reach the server. This step
    confirms that neither stopping nor starting the realm reloads the
    updated manifest.
  ]
• $ ffx component reload /core/ffx-laboratory:echo_realm
• $ ffx component start /core/ffx-laboratory:echo_realm/echo_client
Confirm that the client accessed the server without error. The
manifest was updated and the child server started.

A similar test modifying //examples/components/storage/ shows that
storage survives the `ffx component recreate` command as expected.

No memory or handle leaks. Leak tests: go/component-reload-leak-check

Original-Bug: 66381
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/660862
Original-Revision: 52658c09acfb6af2d8856c75a1b6638fbff6cc87
GitOrigin-RevId: 04ce753a437ed5db32747ba39403baffe6597e63
Change-Id: Ibfda788abc15495e04355cc6d560e5e1cf59fb29