[roll] Roll fuchsia [clippy] Fix lint deduplication
Previously we used the full rendered message to deduplicate lints,
but we found that the note indicating which category implied a lint
only gets added to the FIRST one emitted per clippy invocation.

Example:
```
= note: `--force-warn bare-trait-objects` implied by `--force-warn rust-2021-compatibility`
```

Since we do multiple invocations for toolchains and test variants,
the lint that comes first can change, and will have a different
rendered message content than the same lint in a different invocation,
causing both to show up in `fx clippy` output.

To fix this we're changing the key for deduplication from the rendered
text to a combination of the main message of a lint, it's "code" (e.g.
clippy::bad_bit_mask), and all of its primary spans, which should be
a more consistent indicator of uniqueness.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/681347
Original-Revision: 398eda818b503ddca02d7d7907d1d5ec929e8db0
GitOrigin-RevId: 316cb1d4d88c59b505eb6be02d9cb2ef91920433
Change-Id: I57d075d8e099c577efb43c546908cafe8160c9ae