[roll] Roll fuchsia [wlan] Make sure NetworkPort waits for PortRemoved
The NetworkPort driver component would immediately destruct after
calling RemovePort. However the contract for netdevice requires that
resources associated with the port are not released until the removal
has been completed by a call to Removed. This lead to some rare crashes
when starting and stopping AP interfaces as netdevice would attempt to
make calls to the NetworkPort component after it was destroyed.

Add a completion and wait for it after calling RemovePort so that
Removed can finish and then signal the completion.

Original-Bug: 100725
Test: Included unit test, manually run iperf3 while starting/stopping AP
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/681348
Original-Revision: 7a9180b91e7808ad84b8c7eca4fd2235dcdf9ae6
GitOrigin-RevId: 3d0f6276f52c3c670e1459f5b7137ca73ce77f15
Change-Id: I7799f8b7e97eade83cccccaf5f0f6cd311e01b09