[roll] Roll fuchsia [fidl] Remove usages of coding tables in LLCPP
This CL removes coding tables from the LLCPP type traits and usages of
the coding tables.

However, there are some usages of coding tables that are currently
generated on LLCPP types but are not used in LLCPP. A new trait was
created so the coding tables can still be used in the near term, but do
not increase the size of regular LLCPP usages.

Binary sizes decreased by 300-450k!

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/680463
Original-Revision: de3f79c0d2909b021d7878dfac5d8e19c691295e
GitOrigin-RevId: d18393129d3aa78c12d5a98be6e7ed5baa865695
Change-Id: Idcbc490595ed239432fa800716bb964238468c85