[roll] Roll fuchsia [scrutiny][update-package] Inspect images JSON file
This change modifies how the scrutiny core packages collector
inspects the update package. Previously, the collector looked for
an entry named "zbi.signed" or "zbi" in the update package. Now,
the collector looks for said entry and checks for the same entry
in the "fuchsia" slot of the new images JSON file that is expected
to be included in the update package.

Original-Bug: 98030
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/678358
Original-Revision: 5d03f806b44ac1b9f3e2468575f098b1a72c00c0
GitOrigin-RevId: 70a4e5cc9e478bf31fd11e51bfddc1e195544659
Change-Id: I9a6359151f2d7b350f1e9962757a54106387693a