[roll] Roll fuchsia [netstack3] Include addresses in socketmap tags
This addresses a shortcoming in the original socketmap design. The
assumption there was that all addresses that would conflict would also
have a shadowing relationship in one direction or another. This is not
the case, in part due to the inclusion of device IDs in address vectors.

A UDP socket listening on 0.0.0.0:8080 and bound to device D, for
example, is expected to prevent the binding of another listening socket
on 127.0.0.1:8080 (no device). The address vectors for both these
sockets shadow 0.0.0.0:8080 (no device) but there is no direct shadowing
relationship.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/680114
Original-Revision: 90799b911223143cd96e70a600c88482ac0ca7df
GitOrigin-RevId: b7ed3e8d31683e7570888f43e668607134197dd8
Change-Id: I29f1a3e6b8e65f6813f5ed8c98cae64fbb6e6293