[roll] Roll fuchsia [ffx] Treat "no ssh port" as port 22 for filtering
Changes the logic of the "matching port" subclause of the target
collection matching code to treat 22 as a default for a target when
matching against port 22 is requested.

There's potential here for a larger, deeper, fix to how mdns targets
are added to the collection, since they probably *should* have port 22
specified as their port, but I think it's best to save that for another
day. For now, this enables using `ffx target` commands with a target that
includes the port number when those hosts were discovered rather than
explicitly added and had an ssh port of 22.

Also updated the target fidl documentation to reflect the existing
reality along with these updates. It seems to have been written when
this protocol was much more restrictive.

Original-Fixed: fxbug.dev/98589
Test: fx test host_x64/ffx_daemon_target_lib_test

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/681344
Original-Revision: 1abd5ac3e84d8f18f4d6e1d1a1c7dfebca61d961
GitOrigin-RevId: f8bd6bb164af3e3a6d66cc55cba8fb548737eeb1
Change-Id: I3345ada58a3d9384c44402478301454a4ee589ab