[roll] Roll fuchsia [fdio] Clean and validate paths before processing
This updates fdio entry points that accept a path string to clean and
validate the string before identifying the base target object (the root,
current working directory, or relative directory). This means that paths
of length equal or greater than PATH_MAX are now always rejected.
Previously oversized paths might be accepted if they were prefixed by
some number of leading '/'s. Processing the path in this order also
means that we now consistently first copy the path into a buffer managed
by the fdio_internal::PathBuffer type and then perform processing on
std::string_view objects referencing this buffer. This avoids redundant
strlen() operations on the input and risky pointer chasing and C style
string manipulation routines.

This also updates the internal fdio::open() and
fdio::add_inotify_filter() virtual functions to accept a path by
std::string_view instead of a C style string. These were the last
holdout of C style strings for paths in fdio's internal routines.

This also updates the fdio_chroot test expectations on Fuchsia to match
Linux. Before this change, this test was sending the string
"/../chroot1" to the access() entry point and fdio was passing the
string in non-canonical form to the filesystem server which responded
with ZX_ERR_INVALID_ARGS. With this change, the path is correctly
canonicalized on the client before sending it to the server so the
response is ZX_ERR_NOT_FOUND which translates to ENOENT, matching Linux.

Original-Bug: 98543

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/672089
Original-Revision: 09c1891e0398260153e97ca0edb4726ad81936e1
GitOrigin-RevId: 52e4cb77a9fec73d06a6459ee10aecdf7402ccdc
Change-Id: Ib7a4219232f6ca2429edb0384c548a844a8afa93