[roll] Roll fuchsia [fshost] Fix df for built-in filesystems
df got broken for built-in filesystems (those not mounted using mount).
This CL fixes that and adds a unit test so it shouldn't get broken
again.  This also makes df work when the data filesystem is fxfs which
now runs as a component.

Original-Bug: 99591
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/679186
Original-Revision: 84773b9dc9882ad95dbebc473f17a63624d70948
GitOrigin-RevId: 1bb14b62b28ad6dd8ef42225c75fefb3293a26cc
Change-Id: Ib72d3ebac9ccb36dc8e3977943eb638c4f881e4f