[roll] Roll fuchsia [fdf] Split out driver_host.fidl from fdf FIDL
The fdf FIDL is part of the SDK. The driver host FIDL is only used
internally and therefore not needed in the SDK. This change splits it
out into its own FIDL library and refactors all usages of it. The
DriverStartArgs gets separated out from driver_host.fidl and left in
fdf since it is used in the SDK.

Original-Bug: 97984
Tests: Existing unit and integration tests work. New tests are
not needed as there are no changes in behavior.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/681663
Original-Revision: 51808668c37dd0fd63b3eacbf1db4de05cb3f9f8
GitOrigin-RevId: ecf89c827c43e6efb74e5e4a01488705760acfcb
Change-Id: I2cdd7d80dcb87e61fddad7ac8046b78f74fb2b14