[roll] Roll fuchsia [input-synthesis] Test handling newlines and tabs
I noticed that we never tested input synthesis for nonprintable
chars like \n and \t.  I added this to check that we're doing
the right thing, and it seems that we are.

Added pretty assertions too for easier checking of mismatches.

Then it occurred to me that it would be just fine to merge this.

Original-Bug: 79807

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/680408
Original-Revision: b3b810d9dcd83299410804be4f7890190312df47
GitOrigin-RevId: 86bb4d46ddd7a68712fb125db65f8daaec145adc
Change-Id: I55ddd391525f24453216c9eea1401fb54d8ebe94