[roll] Roll fuchsia [brcmfmac] Synchronously remove netdev port
The network device port was being removed as part of the destruction.
The destruction happens as a result of async remove on the wlanif
device. Unfortunately it seems that new interfaces can be complete
before that destruction has completed successfully. The interface was
destroyed in firmware however so firmware would reuse the interface ID.
When attempting to use that interface ID as the port ID for the
netdevice port the port creation would fail because a port with that ID
already existed, it hadn't been remove completely yet.

Fix this by synchronously removing the port in the call to destroy the
interface, before async remove is called.

Original-Bug: 100854
Test: Manual testing, verify driver handles interfaces up and down
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/681906
Original-Revision: f4200b31e8a78363775afaad0abe97de5e0dec25
GitOrigin-RevId: 3f5fd46faf258ab74e6d3fc4dd4552e134b4ba14
Change-Id: I6737be41a775b075c1f5f3a2884d4707a33cf699