[roll] Roll fuchsia [ui-test-manager] Instantiate base realm using RealmBuilder::CreateFromRelativeUrl().
RealmBuilder recently added a feature to enable users to create a realm
with a topology specified in .cml. Note that this is different from what
ui test manager does today: we currently create a monolithic component
to represent the entire UI topology and add it as a child of the test
realm. With CreateFromRelativeUrl(), there is no monolithic "ui realm
component". Instead, the ui components (scenic, RP/SM, etc.) are direct
children of the test realm, and can be referenced by name. This change
has several benefits:

(1) It prevents false cyclic dependencies with the monolithic "ui realm
component". For example, I wasn't able to write a test for the a11y view
with flatland scene manager, because the a11y manager needed to consume
flatland services from the "ui realm component" AND serve
fuchsia.accessibility.scene.Provider to the "ui realm component".
This cycle represents an invalid component topology. With the new
approach, we don't have the "ui realm component" as a bottleneck for
services going into/out of the ui stack.

(2) It simplifies some of the service routing in UITestManager, because
we can specify which services to expose out of the test realm directly
in .cml.

(3) It integrates better with static configs. With the "ui realm
component", we had no way to override configs for ui components, because
they weren't direct children of the test realm root.

Test: semantics-integration-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/677662
Original-Revision: d8e5dfd7a52ad4bb1bea4682f88ac5a887acea3f
GitOrigin-RevId: 3be6c6e7856a61065de5c936fb3114af0b002cfa
Change-Id: I0fcf76e979df00929958e1821cd466937359dcf5