[roll] Roll fuchsia [fidl] Remove clang-format column limit for fidlgen
To allow our fidlgen templates to be optimized for maintainance rather
than generating readable output we pass generated files through a
language-specifc formatter. For C++ this is clang-format. Unfortunately
clang-format can be slow for large, complex files.

It turns out that one way to mitigate this is to remove the column
limit, passing "ColumnLimit: 0" as a configuration option.
Counterintuitively this doesn't actually result in crazy long lines,
just arguably mildly worse line breaking choices, though that's
subjective. Look at the changed goldens in this commit to see the impact
on the output.

The upside is a performance win. On my workstation, the time to format
all of the goldens drops from 15s to 5s with a modest drop in maximum
memory usage from 72M to 68M.

Formatting the LLCPP bindings for fuchsia.posix.socket - our largest
generated code (that is excluded from formatting for performance
reasons) runtime drops from 41s to 21s but max RSS for both was about
1.2G.

I don't think this will let us stop using a size limit for formatting
but at least it should speed things up

The clang-format args used to be scattered across a Go file and some
BUILD.gn files but now they're just all in the generator.go, making it
simple to experiment with tweaks in the future.

Original-Bug: 78303
Original-Bug: 49757
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/680908
Original-Revision: 9e75c206ce23aab6d62d03ae5830610b4384b3b6
GitOrigin-RevId: 4c375ef704e8d3e198f16d4454ed91e8ce6580da
Change-Id: Ied36a2b1a78774a9cce14b8bc0d75a06cd802321