[roll] Roll fuchsia [docs][lsan] Add a short description of how LeakSanitizer works
I thought this would be useful to add, because otherwise the
information is three steps away from this page, via these links:

https://clang.llvm.org/docs/LeakSanitizer.html ->
https://github.com/google/sanitizers/wiki/AddressSanitizerLeakSanitizer ->
https://github.com/google/sanitizers/wiki/AddressSanitizerLeakSanitizerDesignDocument

Some other leak checker tools have relied on programs freeing
allocations before running the leak checker and exiting.  It's easy
for there to be misconceptions that LeakSanitizer has similar
requirements, and that allocations that are never freed (or threads
that never exit) are considered leaks, which is not necessarily the
case with LeakSanitizer.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/679995
Original-Revision: 6035045037b8ddfddf49ed12cbffff69a34cb257
GitOrigin-RevId: 6f6bb1f88dee0cbfb5dc19a29b4cafaece29bd8d
Change-Id: Idc27989ee7fbd3e7555e9428a3b23522d3fbbe22