[roll] Roll fuchsia [audio][clock] Add RecoveredClock
This CL adds a class which can be used to recover a clock from a stream
of position updates (e.g. from a device driver). This is being factored
out of AudioDriver and AudioClock. The behavior should be identical to a
a DeviceAdjustable AudioClock as used in the following code:

https://cs.opensource.google/fuchsia/fuchsia/+/main:src/media/audio/audio_core/audio_driver.cc;drc=73cfbf4538bbcec51bf87420803ca2da65060f8d;l=588
https://cs.opensource.google/fuchsia/fuchsia/+/main:src/media/audio/lib/clock/audio_clock.cc;drc=a05c6d9f6f02df7d46611e8b2626e9ad96cd4a1f;l=318

Original-Bug: 87651
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/680506
Original-Revision: e94ab18ea68dc9f7d243eb11496ea8aaa9b9a3e8
GitOrigin-RevId: 4f41b3a47253d6ff51b144c2063ba207e7dea644
Change-Id: I668ecd3b0abf7de593c69c6dc5d80b718ea9d025