[roll] Roll fuchsia [fxfs] Fix tracking of mark_for_deletion bytes.
The mark_for_deletion call is used to tag encrypted ObjectStore data
for deletion without having to explicitly delete individual extents.
It issues a `MarkForDeletion` mutation which is committed as with any
other mutation to the Journal.

`MarkForDeletion` is applied differently for live and replay scenarios.

In live scenarios, we must consider bytes committed for deletion but
cannot reuse them until after a device flush without risking corruption
(premature extent reuse). After a device flush, we can reuse any extents
associated with `MarkForDeletion` mutations that were flushed. (This
is why we must track the journal offset of each of these mutations at
commit time.)

In replay scenarios, we know it *must* be applied (as it's already on
disk) and can reuse the extents immediately.

The actual marking is done via a HashSet stored in AllocatorInfo
metadata.  All owner_object_id in this set are filtered out of
iterators used for allocating new extents.

Finally, when the allocator itself is flushed and layers are compacted,
we apply the same filter (to produce a file free of marked extents)
and clear the HashSet in AllocatorInfo before it's written to disk.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/679264
Original-Revision: 402fbb368fbac94778941cb84f4ee76ebeb8940f
GitOrigin-RevId: 54c73ee839401d1f26cc847ed02352e66be9854e
Change-Id: I193e907ee59863bb469930a73187e069ff410c50