[roll] Roll fuchsia [component_manager][storage] Introduce Storage Persistence
Introduce the boolean field `persistent_storage` to collection
declarations. When this field is set on a collection, the isolated
storage used by the collection's descendants continues to exist after
the component is shutdown or destroyed (Note: static descendants have
implied storage persistence because static components don't get
destroyed).

Struct `ComponentInstance` stores a `persistent_storage` field. When a
new child component instance is created, is `persistent_storage` setting
is set from the component's collection decl, if any. If a collection
decl doesn't exist, the `persistent_storage` value for the component
instance is inherited from its parent component.

When a component instance is destroyed (via
component.destroy_instance()), the `persistent_storage` field is checked
and if set to true, the component's storage is not deleted.

The persistent storage can be destroyed directly through a
StorageAdminRequest::DeleteComponentStorage request.

The reason for this change is to provide a mechanism for the storage
data of dynamic component instances to outlive the lifetime of the
instance itself, such that when the next instantiation of the component
is started, it can access the storage data of its predecessor.

Original-Bug: 92479
Multiply: fuchsia-pkg://fuchsia.com/component-manager-tests#meta/component_manager_tests.cm:100
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/679904
Original-Revision: 1938333349db80d03e5aac2ec22eda09bb925037
GitOrigin-RevId: c825dab051988981d5864d91b0e3e5cf3b365eed
Change-Id: I2279b26a7b99b4de104e2ee53e1c71064b9e4d6b