[roll] Roll fuchsia [netstack3] Bind unbound socket on sendto() call
Handle an application calling sendto() on an unbound socket by first
binding it to the unspecified address and then sending a packet on it.
This behavior is consistent with Linux, as determined through
experimentation: if the send fails, i.e. because the address is
unreachable, the socket is still bound to a local address afterwards.

Original-Bug: 100939
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/680943
Original-Revision: 8bc11de1b78e28547759ede549a1856ab667ffa4
GitOrigin-RevId: c9c53114d6da67a936113df2ceeff65e79953f31
Change-Id: I42829ae00846dce5c25b41a392e85fd35df03362