[roll] Roll fuchsia [fidl][gidl] Test handle closure during failures
This tests a case that was encountered when building the new LLCPP
encoder and decoder. It is important to ensure the handles are
closed regardless of the location of the error and that the
resulting handle array is properly sized.

Original-Bug: 100931
Original-Bug: 96626
Original-Bug: 100945
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/681788
Original-Revision: a03fe93900043018b8e44fbe226d34ac71a9566f
GitOrigin-RevId: 94709c4f9fdf1b7a119a44362b1e05daebf2b086
Change-Id: I06c1fcafd607d34b64a24926aa54ac885f6791dd