[roll] Roll fuchsia [artifactory] Stop uploading Jiri snapshot
The snapshot is already uploaded by the recipe, which is the more
"proper" place to be doing this (during the checkout phase), not
post-build.

The reason this was added was because users had requested that the
snapshot show up in fuchsia-artifacts buckets, but fxbug.dev/51741
migrated the Jiri snapshot upload to fuchsia-artifacts, so this is
always a no-op for artifactory anyway. (It just gets deduplicated every
time).

Original-Bug: 51741
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/682543
Original-Revision: cc55bfe3bb883930f56cd159e8b19cc699723bac
GitOrigin-RevId: 73c3df8887619f4ec3e9e41ebb7e361383d37b25
Change-Id: I2ca05cf469723a110cdc386976e57f40ac095c9d