[roll] Roll fuchsia [artifactory] Move nonexistence checks earlier in execution
Artifactory already filters out nonexistent files, but only did this at
upload time. This change modifies the execution to filter prior to
emitting the manifest, so that the manifest is not polluted with files
that do not exist.

This allows gcs-util to enforce that every file in the input manifest
does exist, which is a good assertion to have. Each project should take
responsibility over emitting a manifest with entirely valid references.

Original-Bug: 92697
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/682802
Original-Revision: b56de5035dc537abeee00f3536f0867c5078c3b2
GitOrigin-RevId: b45eb118220405cb6024e9769d2b6a72e583f324
Change-Id: I9b45574b74d4c24ce647a6f180941bc20da32172