[roll] Roll fuchsia [component_manager] Purge invokes Destroy
This change causes the Purge action to automatically invoke Destroy.
Component instance destruction paths are changed to schedule the Purge
action and synchronously wait on it.

This is the minimal change to get the behavior proposed in
fxbug.dev/100652. It is preparation toward merging replacing Purge
entirely with Destroy, which will happen in a followup.

Original-Bug: 100652, 89503

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/677672
Original-Revision: b5ad79b911e1f949c86d1bfe3345a2022a1c62f4
GitOrigin-RevId: 676b08981bb385acf147597195c7724812cdf27f
Change-Id: Id61d8b10c45054e6c3587ee5bf87a4b64bf595cf