[roll] Roll fuchsia [build] Add compatibility tests for past versions
Tracks both the current and past supported API levels
in platform_version.json. The set of supported levels
is seeded with [7] for the time being, but should be
updated to include 8 when we are ready to enforce
compatibility for that level.

Original-Bug: 92348
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/678857
Original-Revision: 842367671c1e0bb2cdbb9bd2c3f1c57c8fa9190c
GitOrigin-RevId: 414cb90ac4fb238a819e4c1141686b5ef7faf5a1
Change-Id: Ifd196ca83c3a25f2ae78099efe3a886f7545281e