[roll] Roll fuchsia [bt][core] Remove unused production v1 components
bt-gap and bt-init are now CFv2 in production, so we can remove the v1
components. For bt-gap, this just means removing the v1 manifest and
build targets. For bt-init, the v1 component was actually a separate
binary with a different main.rs, so remove that as well as the manifest/
build targets.

Original-Bug: 89297
Test: `fx test //src/connectivity/bluetooth` to verify nothing broke.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/680571
Original-Revision: 73c83a856ee4da72e7d7732e79a4cba85a88729b
GitOrigin-RevId: 9262d5d671eaed14963ccfb20074a781f3ec627b
Change-Id: Ia1bf936062b09a085c50255dc87a46a102a145b6