[roll] Roll fuchsia [test_manager] Creates a specific type of test for starnix
The starnix test currently run a linux binary with a starnix runner
outside of the test realm.

This is problematic for 2 reasons:
- The test is not hermetic at all.
- There is no code coverage on the runner code

To fix this, the runner will instead be run inside the realm of the
test.

The runner requires a number of permission that we do not want to give
to every test. To allow only giving these permissions to the test that
needs it, this CL introduce a specific type of tests and a specific
collection to run these tests in.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/682964
Original-Revision: 43a6feb05c66bfd50f2184ba0a529cbac26deb19
GitOrigin-RevId: 2d68edfb18632605cf06a2fabfed2e34bd93c228
Change-Id: I7801db3c28c427a1b670b6eac4e6c3a87e31adc1