[roll] Roll fuchsia [testrunner] Move test outputs after recording test result.
In http://fxr/680306, we used a temp dir for the test outputs dir which
some tests write to, and then moved those contents back to the
originally designated output dir. This broke tests which use the
testparser to store paths to those outputs as outputfiles of the test in
the test result because the paths would point to the tmp dir that no
longer existed. This change allows testparser to access the paths in the
tmp dir and record the output files which will then be moved (and paths
updated) in the call to outputs.Record(). After the output files are no
longer pointing to paths in the tmp dir, we can then move the rest of
the contents to the original test output directory so that they can be
uploaded with the swarming task outputs.

Original-Bug: 100948
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/682023
Original-Revision: 785ba182692dc2a82e23c898142aae8c6084b8c4
GitOrigin-RevId: 237f8ddfc3c9fc145b3a3c1b55ae3bf6ad3f2621
Change-Id: I1177afc09749dcb9878c2d63ee44567a2e3380ea