[roll] Roll fuchsia [bt][fastpair] Use ECDH to compute a shared secret
The anti-spoofing AES key is used in the Fast Pair key-based pairing
procedure to verify the identity of both parties. It is calculated by
taking the remote device's Public Key (An (x,y) point on the
elliptic-curve), the local device's Private Key (An integer on the
elliptic-curve), and applying ECDH.
See the GFPS specification for more details:
https://developers.google.com/nearby/fast-pair/specifications/service/gatt#procedure

Define crate-local types representing Public/Private keys to be used in
the pairing procedure.

Use the P256 and elliptic-curve crates to calculate this shared secret.

Original-Bug: 96222
Test: fx test bt-fastpair-provider-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/672755
Original-Revision: 88283ad1da18dcabb6141f88e0ff68ceeb13a3aa
GitOrigin-RevId: e2d4d8886e7233b1aabe159938d206737324c557
Change-Id: I8543dad0a62c0d48eab95e0ba8e186a53bdc599e