[roll] Roll fuchsia [kernel][mutex] Early terminate the spin-phase
Early terminate the spin-phase of the kernel's adaptive mutex if there
is a reasonable chance that the thread who owns the mutex has the same
CPU assigned to it as the thread which is currently attempting to
acquire the mutex. Frequently, this would be because a deadline thread
has preempted a fair thread who is holding a mutex that the deadline
thread wants.  Spinning would result in consuming precious time from
the deadline schedule, and we would prefer to drop out of the spin and
move into the block in order to make profile inheritance happen.

See the comment near the top of Mutex::AcquireContendedMutex, however.
This is a "best effort" heuristic.  It is not easy to know which CPU
the owner thread is actually currently assigned to, and this approach
can cause a thread to drop out of a spin when it normally might not,
or to continue to spin even when it probably shouldn't.

There exist ways to reduce the potential for getting our guess wrong,
but they are more expensive than this simple approach owning to:

1) Extra cycles spent on the fast of lock acquisition and release.
2) Extra cycles spent in several scheduler operations where a list of
   mutexes need to be updated because a thread blocks or is assigned
   to a different CPU.
3) Extra storage in both the Thread and Mutex objects in order to
   give Threads the ability to maintain a list of currently owned
   Mutexes.
4) Significant extra complexity in the form of needing to make sure
   that everything remains thread safe and has proper memory ordering
   semantics.  Things which would typically be enforced using some
   form of mutex or spin-lock, but which is not an option here because
   of the nature of the problem.

Original-Bug: 100122

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/681804
Original-Revision: 689b2bffae4d784172e496cad321ee8f9f660191
GitOrigin-RevId: 7902b32f6e756dca39177e7557b14dad80bf246e
Change-Id: I6f4ed239626bc35be16c87781c42b744df8e958d