[roll] Roll infra/recipes [btu] Allow dest="." in CIPD upload manifests
This is a convenience to allow projects to upload the contents of an
entire directory as a CIPD package. It will be useful for uploading
things like debug symbols, where the desired CIPD package contents are
often already in a single subdirectory of the build directory that
should be uploaded as-is.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/683262
Original-Revision: 8380ba7ccd5c92890f3061c7825325f09d67944a
GitOrigin-RevId: 4aabbc6be4289f4658d73418698cc1a526361591
Change-Id: I2e13c3cfa2b76f8524d0e125bcc62d53f2163731