[roll] Roll fuchsia [starnix] Run starnix runner inside starnix tests
Until now, when running linux tests, the starnix runner was run from the
system.

This was problematic for hermecity, as well as preventing to get code
coverage.

This update the starnix test runner so that the starnix runner is now
run inside the realm of the test.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/682902
Original-Revision: ca4d988b7f8c4414287de11456c2c9b351a5fb1e
GitOrigin-RevId: 490e6de4402090932221310e064ed3f535a5de10
Change-Id: Ib3e540c2e64cf232b9af0a5b78e9fcfaf8f13b0b