[roll] Roll fuchsia [build] Delete sdk_package
This template is broken because it attempts to use fuchsia_package() on
the host toolchain. fuchsia_package() cannot be used with anythign but
the target toolchain, and when it does it produces an empty group.
Delete this template and all of its uses, which were all unfortunately
no-ops.

This was written and reviewed by Misha and Sean, both of whom have
unfortunately left the team.

Original-Bug: 100612
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/680933
Original-Revision: 012afda6fbe9e15a618ba110f024196c4ff3b36e
GitOrigin-RevId: c4d9960d0f3aaeb438efc8d2cf4abdf5de461922
Change-Id: Idd4fe459dd79de2370fbb1c78120504f0f7855fe