[roll] Roll fuchsia [installer] Log FVM write status.
The FVM write takes a long time. In some instances the installer tests
flake, and it's not clear if it's hanging or just taking a while to
write the FVM.

This CL adds some progress logging so we can get an idea of how
quickly/slowly the installer is progressing during the FVM write.

Original-Bug: 100712
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/683363
Original-Revision: f43f9b45dd36a1070e8264525c268b8ac48a0a9e
GitOrigin-RevId: 5e1310d6aa595c8c64dfcab839ec8a86720388a1
Change-Id: I0fdf32a75884bc2d0a059e682c7271db2c54ba6d