[roll] Roll infra/recipes [owners] Fix ownership of expectation files
(Fix for fxrev.dev/683263)

Based on the syntax documentation for owners files, globs must match an
entire suffix of the file path in order to apply to a file. So
`*.expected` isn't sufficient to match all expectation files, we need to
make it `*.expected/*.json` so that it matches a full suffix of each
expectation file path.

Likewise, replace * (which can't match slashes) with ** (which can match
slashes) in toolchain recipe globs to ensure the globs match the
contents of resource directories.

I tested this by uploading to a sandbox branch and using the "list code
owners for path in branch" via gob-curl to validate that the code-owners
plugin returns `"owned_by_all_users": true` for expectation files, and
includes the appropriate toolchain team owners for toolchain recipe
resource files.

https://android-review.googlesource.com/plugins/code-owners/Documentation/rest-api.html#check-code-owner-config-files

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/683266
Original-Revision: 366287042f9a1e02cee7285ce4a7a96d0ad4b516
GitOrigin-RevId: f7ce2ae65a0930c1d929b34aba0aac5d960eb8e3
Change-Id: I88aba017f8d992fa84266be4b6bee31f11db8f57