[roll] Roll fuchsia [cleanup] remove all stale suppressions in zircon
bug: 58162

This change is part of the C++ implicit conversions project. These are the build files that compile without more changes. The purpose is to shorten the list of stale suppressions in the kernel to find which ones still need to be fixed.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/683422
Original-Revision: c4b4f21957bc316e5eaa3d4ca76f4e67cb8e241e
GitOrigin-RevId: 041de85babb1cb6cdc1dc9231df98213c5e5c9ae
Change-Id: I8398dd59f1aba9df0f2468f9601a4fdd9fc2b77a