[roll] Roll fuchsia [codec][vaapi] Fix buffer size assert when fuzzing
The fuzzer code was allocating output buffers without knowing the size the video
stream requested. When the decoder tried to write to a buffer this could
cause an assert if the buffer isn't big enough.

To fix this, allocate buffers in
onCoreCodecMidStreamOutputConstraintsChange so that we know the output
constraints and the expected sizes of the buffers.

Since allocating buffers can take a while, we need to ensure the fuzzing
doesn't stop prematurely before it happens. To do this we can queue an
end of stream after the video data, and wait for the callback to be
received before exiting the fuzzer. Since some crashes may only happen
on unclean teardowns, half of invocations will avoid sending end of
stream and will be terminated like before.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/682664
Original-Revision: 634dc4d9e6bdd4ce547b3ba339682d4f3f1eee25
GitOrigin-RevId: a1ad1c2cc887bb0176135c40da65f967b5e9cb12
Change-Id: Ic8116d110e425b42073d5404ad8ef94fafcd5bc3