[roll] Roll fuchsia [audio][drivers][configurator] Add DAI interfacing
We add the infrastructure to talk to the DAI drivers
in future CLs. This includes the DAI methods handling
support for configurator, the ability to get a proxy
and then connect to devices in /dev/class/dai, and a
realm test driver.

Original-Bug: 82451
Test: configurator-unittests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/682344
Original-Revision: 302bcbd05665e6451dad4675bdc06935e31b16d3
GitOrigin-RevId: b78653350df8fa6e8568f475beda007a87d19c74
Change-Id: I6fd9a1e71cd6342204215a368fd6f765471ff569