[roll] Roll fuchsia [sdk] Remove support for fsys2.Events from Dart Realm Builder
Dart Realm Builder will be published to the SDK, but Events (under
fuchsia.fsys2) are not supported in the SDK yet, so they must be
removed. Events are coming to the Fuchsia SDK in the near future, and
this functionality can be restored at that time.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/683851
Original-Revision: b1c5124a07ad75e11064c4ee3ab416f9a52966a7
GitOrigin-RevId: 3a5b60c1fae0d73d4dd600289d6474a7d60947d2
Change-Id: Ie02722c58840e14ce3abd7451f807fc2a878f4e8