[roll] Roll infra/recipes [artifacts] Remove api.artifacts.upload, use artifactory JSON output
In favor of api.artifacts.upload_from_manifest. Uploading is now a
two-step process: one, the project produces the upload manifest. Two,
the infrastructure reads the upload manifest and executes the uploads.

This swaps artifactory to use the "upload manifest" flow, which is part
of a soft-transition to consolidate GCS upload logic as a common feature
to be used by all projects in-tree and out-of-tree.

Original-Bug: 92697
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/682204
Original-Revision: 61e2f9fc98c04b45b3034051b313dbc8ff9ffaa1
GitOrigin-RevId: 2d4e5d8d66937936af0f82770df6c8f2b28bc059
Change-Id: I65c5c67d83cadcbe3aa45c4290f7fffbe72bcd6c