[roll] Roll fuchsia [bt][hci] Don't notify canceled SequentialCommandRunner command callbacks
Don't notify SequentialCommandRunner command callbacks that were in
flight after Cancel() has been called, a command fails, or the
SequentialCommandRunner is destroyed. This was always the case for
commands which were in the queue but not yet sent to the command
channel, but this is now also true for commands which have already
been sent to the command channel, but not yet completed.

Original-Bug: 100658
Test: //src/connectivity/bluetooth/core/bt-host

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/681447
Original-Revision: 032541a6db478ee42dd58c169f11b3969a02d138
GitOrigin-RevId: 4d44f2fcf20a8297364931597332544bfd0b44c5
Change-Id: Id22786ce1b3b12e7dac97c93ee1ee52b11f2e5c6