[roll] Roll infra/recipes [edk2] Initial commit
This is a new recipe for building EDK2 UEFI firmware for QEMU for all
architectures. It supersedes the OVMF recipe (which just handled the x86
side of things). There is still a bit to do in this recipe before it is
ready for primetime. However, it is useful in its current form as it can
be used to have the x86 and arm64 firmware built and uploaded to CAS for
local download and iteration with the firmware itself.

Original-Bug: 99325
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/675683
Original-Revision: 5cd0bde3882e1d4abe7c25a70c35186b97e6a324
GitOrigin-RevId: b31aea2e0d8ab6f2d4a9dc5985a86467ef9444c3
Change-Id: I07bc65fd2e421e14bdef0da8c1e847a096d4fe47