[roll] Roll fuchsia [audio][drivers] Fix audio-driver-ctl connection check
Rely only on not being able to communicate with a driver
(getting formats) to find out if a driver is present at
a particular devfs location. This because we use a feed
forward API which may not return an error when setting up
a connection.
fxrev.dev/678038 checked for stream channel peer closed
which fails depending on timing since it takes some time
for the signal to get asserted, removed the bad check.

Test: audio-driver-ctl info on Nelson shows driver info

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/684163
Original-Revision: f59eddb3fb98118f259ea4c232e7c73ca299849e
GitOrigin-RevId: 218af8424626fc2f53917cb8263a646b08f7cc45
Change-Id: Ied74c42bae25afc90276f909ff96effb97aa7020