[roll] Roll fuchsia [rust] Use the path from the checkout for llvm-profdata-rust.
The rust version of llvm-profdata links LLVM as a shared library so we'd
need the shared library file together with the llvm-profdata binary in
order to make the tool work. We can possibly upload both with the
orchestration inputs so that the coverage processor builder can use it,
but that would require some more refactoring to be able to preserve the
relative paths and somehow know the path to the llvm-profdata tool
within the rust prebuilts dir. On the other hand, the coverage processor
builder already does a checkout from which it could access the tool
directly, getting the path through the tool_paths.json.

Original-Bug: 95345
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/683869
Original-Revision: 49f3127275c1a00ab8c71c29692f66733df1667b
GitOrigin-RevId: a945e4807a001426d684383b49e2ec785c96d3a8
Change-Id: I569fc239e83fde36a9f520e6fc5f79126917319e