[roll] Roll fuchsia [audio_core] Reboot audio_core on terminate
A version of audio_core built out-of-tree is currently labeled
reboot-on-terminate. The original motivation was that some code
(including in-tree code) is (currently) not designed to correctly
reconnect to audio_core and reconstruct audio state after an audio_core
crash, so it was deemed safer to just reboot the system.

The in-tree audio_core was never labeled reboot-on-terminate, which was
simply an oversight. This CL corrects that oversight.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/683883
Original-Revision: dcae2f214b017271e5fbe4ebcd9518a14a1715c2
GitOrigin-RevId: 6dff93e2e482c67c2899f466329eaa39d49e1092
Change-Id: Iec0947f4b720e8af03af28a5735245f6472c326b