[roll] Roll fuchsia [libabr] fix magic constant definition
Putting a static const variable in a header actually tells each
translation unit to make its own copy of the variable, rather than
sharing a single copy.

This can cause a (small) duplication, but more importantly, files
which include this header but *don't* use this particular constant
will trigger an unused variable compiler error.

This CL just switches it to be a macro instead. The "right" way to do
this is to define an extern char* in the header and provide the
definition in a .c file, but since we also want to declare the length
constant immediately after it's best to keep the definitions
side-by-side.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/684362
Original-Revision: 0e37651f9c9e1dbd7378681aab4a5bf1dda7b334
GitOrigin-RevId: 7a62a24abe0a7986f6b2dd59c43cdb6872db5657
Change-Id: I3f7ab04abef454d7feddcf376c1f20453af77c7a