[roll] Roll fuchsia [fshost][test] Run fshost tests with Fxfs
Templatizes the fshost integration test suite so we can run the test
with various fshost configurations.  This permits running the test suite
with Fxfs, which is also added by this CL.  (f2fs can be added in a
follow-up; it was not turnkey so I left it out of this CL.)

A previous attempt to run the fshost integration tests with varied
configuration was fxrev.dev/679445, which used Gtest parametrized tests
and RealmBuilder, did not allow variance in the component manifest
files, which would cause problems in the future.  This approach is less
tidy but allows both the structured config and the component manifest to
be varied.

Original-Fixed: 99714
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/683026
Original-Revision: 10a12dd41f08bcaaeb1ae1f0788f8278baaa2dc0
GitOrigin-RevId: 5d772897fcdde554d8a838712333667407e546a4
Change-Id: I512b168311da5a543b4b33a2bf06308dfdb361a1