[roll] Roll fuchsia [pkgctl] Add newline to `pkgctl get-hash` output.
This makes the pkgctl command line output for get-hash consistent with
the rest of the tooling. Previously, the newline was missing for the
get-hash subcommand, causing the command line prompt to appear directly
after the output hash instead of the next line.

Original-Fixed: 95157
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/662665
Original-Revision: 16a008c7869a62b3e70b51da4151dfdf57800896
GitOrigin-RevId: 4f792b54bf7516983ac52f87d4cc9e0572878bb7
Change-Id: I4d8c009a1d65840ce5a9eedbd553efd0ae2b6e0a